ROPECA contrast analysis facade for nested input

ROPECA contrast analysis facade for nested input

ROPECA contrast analysis facade for nested input

Value

An R6 class generator.

Details

Encapsulates the pipeline: strategy_lm -> build_model -> Contrasts -> ContrastsROPECA.

ROPECA fits per-peptide linear models and aggregates peptide-level p-values to the protein level. Input must be nested LFQData (peptide subject_id strictly below the protein hierarchy).

See also

Super class

prolfqua::ContrastsInterface -> ContrastsROPECANestedFacade

Public fields

model

Model object (peptide-level)

contrast

ContrastsROPECA object

.lfqdata

stored reference to input LFQData

.contrast_names

names of the requested contrasts

Methods

Public methods

ContrastsROPECANestedFacade$new()

ContrastsROPECANestedFacade$get_contrasts()

ContrastsROPECANestedFacade$get_missing()

ContrastsROPECANestedFacade$get_Plotter()

ContrastsROPECANestedFacade$to_wide()

ContrastsROPECANestedFacade$clone()

Inherited methods

prolfqua::ContrastsInterface$column_description()

prolfqua::ContrastsInterface$contrast_summary_table()

prolfqua::ContrastsInterface$extra_artifacts()

prolfqua::ContrastsInterface$filter_significant()

prolfqua::ContrastsInterface$get_config()

prolfqua::ContrastsInterface$get_contrast_sides()

prolfqua::ContrastsInterface$get_ora()

prolfqua::ContrastsInterface$get_rank()

Method new()

initialize

Usage

ContrastsROPECANestedFacade$new(lfqdata, modelstr, contrasts, ...)

Arguments

lfqdata

nested LFQData (peptide-level)

modelstr

model formula string (e.g. "~ group_")

contrasts

named character vector of contrasts

...

passed to strategy_lm

Method get_contrasts()

Get contrast results with standardized column names.

ROPECA's beta.based.significance is mapped to p.value and FDR.beta.based.significance to FDR.

Columns not directly produced by ROPECA are derived heuristically:

std.error = diff / statistic (algebraic: t = estimate / SE)

sigma = mad.estimate (MAD of peptide fold changes)

df = n_not_na (number of contributing peptides)

conf.low/conf.high via diff ± qt(0.975, df) * |std.error|

Usage

ContrastsROPECANestedFacade$get_contrasts()

Method get_missing()

get protein x contrast pairs that could not be estimated

Usage

ContrastsROPECANestedFacade$get_missing()

Method get_Plotter()

get ContrastsPlotter (uses standardized column names)

Usage

ContrastsROPECANestedFacade$get_Plotter(fc_threshold = 2, fdr_threshold = 0.1)

Arguments

fc_threshold

fold change threshold

fdr_threshold

FDR threshold

Method to_wide()

convert results to wide format

Usage

ContrastsROPECANestedFacade$to_wide(columns = c("p.value", "FDR", "statistic"))

Arguments

columns

value columns to pivot, default c("p.value", "FDR", "statistic")

Method clone()

The objects of this class are cloneable with this method.

Usage

ContrastsROPECANestedFacade$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

istar <- sim_lfq_data_peptide_config()
#> creating sampleName from file_name column
#> completing cases
#> completing cases done
#> setup done
lfqdata <- LFQData$new(istar$data, istar$config)
lfqdata <- lfqdata$get_Transformer()$log2()$lfq
#> Column added : log2_abundance
contrasts <- c("A_vs_Ctrl" = "group_A - group_Ctrl")
fa <- ContrastsROPECANestedFacade$new(lfqdata, "~ group_", contrasts)
head(fa$get_contrasts())
#> determine linear functions:
#> get_contrasts -> contrasts_linfct
#> contrasts_linfct
#> Joining with `by = join_by(protein_Id, peptide_Id, contrast)`
#> # A tibble: 6 × 14
#> # Groups:   contrast [1]
#>   facade        protein_Id  modelName contrast  avgAbd    diff     FDR statistic
#>   <chr>         <chr>       <chr>     <chr>      <dbl>   <dbl>   <dbl>     <dbl>
#> 1 ropeca_nested 0EfVhX~0087 ROPECA    A_vs_Ctrl   4.27 -0.0742 5.28e-2     -1.75
#> 2 ropeca_nested 7cbcrd~5725 ROPECA    A_vs_Ctrl   4.51  0.741  9.91e-5      8.79
#> 3 ropeca_nested 9VUkAq~4703 ROPECA    A_vs_Ctrl   4.47 -0.598  6.91e-6    -12.7 
#> 4 ropeca_nested BEJI92~5282 ROPECA    A_vs_Ctrl   4.23  0.277  1.87e-3      3.94
#> 5 ropeca_nested CGzoYe~2147 ROPECA    A_vs_Ctrl   4.76 -0.310  3.74e-5     -9.26
#> 6 ropeca_nested DoWup2~5896 ROPECA    A_vs_Ctrl   4.43  0.295  1.38e-6     14.7 
#> # ℹ 6 more variables: std.error <dbl>, df <int>, p.value <dbl>, conf.low <dbl>,
#> #   conf.high <dbl>, sigma <dbl>
fa$to_wide()
#> # A tibble: 10 × 5
#>    protein_Id diff.A_vs_Ctrl p.value.A_vs_Ctrl FDR.A_vs_Ctrl statistic.A_vs_Ctrl
#>    <chr>               <dbl>             <dbl>         <dbl>               <dbl>
#>  1 0EfVhX~00…        -0.0742       0.0423         0.0528                   -1.75
#>  2 7cbcrd~57…         0.741        0.0000496      0.0000991                 8.79
#>  3 9VUkAq~47…        -0.598        0.00000138     0.00000691              -12.7 
#>  4 BEJI92~52…         0.277        0.00112        0.00187                   3.94
#>  5 CGzoYe~21…        -0.310        0.0000150      0.0000374                -9.26
#>  6 DoWup2~58…         0.295        0.000000138    0.00000138               14.7 
#>  7 Fl4JiV~86…         0.0700       0.740          0.823                     2.08
#>  8 HvIpHG~90…        -0.384        0.00000389     0.0000130                -7.28
#>  9 JcKVfU~96…        -0.0634       0.00218        0.00312                  -1.87
#> 10 SGIVBl~57…        -0.122        0.993          0.993                    -3.88