Compute contrasts with group mean imputation

Compute contrasts with group mean imputation

Compute contrasts with group mean imputation

Details

If there are no observations in one of the groups for some of the proteins, the group mean cannot be estimated. Therefore, assuming that the observation is missing because the protein abundance is below the detection limit, we substitute the unobserved group with the median of protein abundances observed only in one sample of the group. The variance of a protein is estimated using the pooled variance of all observations of all groups.

See also

Super class

prolfqua::ContrastsInterface -> ContrastsSimpleImpute

Public fields

subject_Id

subject_id e.g. protein_ID column

contrasts

array with contrasts (see example)

modelName

model name

contrast_result

data frame with results of contrast computation

lfqdata

data frame

confint

confidence interval

p.adjust

function to adjust p-values

global

Take global or local values for imputation

present

default 1, presence in interaction to infer limit of detection.

Methods

Public methods

ContrastsSimpleImpute$new()

ContrastsSimpleImpute$get_contrast_sides()

ContrastsSimpleImpute$get_contrasts()

ContrastsSimpleImpute$get_Plotter()

ContrastsSimpleImpute$to_wide()

ContrastsSimpleImpute$clone()

Inherited methods

prolfqua::ContrastsInterface$column_description()

Method new()

initialize

Usage

ContrastsSimpleImpute$new(
  lfqdata,
  contrasts,
  confint = 0.95,
  p.adjust = prolfqua::adjust_p_values,
  modelName = "groupAverage",
  method = "V1",
  global = TRUE,
  present = 1,
  minsd = 1
)

Arguments

lfqdata

LFQData

contrasts

array of contrasts (see example)

confint

confidence interval

p.adjust

method for p-value adjustment - default Benjamini Hochberg

modelName

default "groupAverage"

method

internal default V1

global

default TRUE use all or per condition data to impute from

present

in at most how many samples the protein should be observed

Method get_contrast_sides()

get contrasts sides

Usage

ContrastsSimpleImpute$get_contrast_sides()

Method get_contrasts()

table with results of contrast computation

Usage

ContrastsSimpleImpute$get_contrasts(all = FALSE)

Arguments

all

FALSE, do not show all columns (default)

Method get_Plotter()

get ContrastsPlotter

Usage

ContrastsSimpleImpute$get_Plotter()

Returns

Contrast_Plotter

Method to_wide()

convert contrast results to wide format

Usage

ContrastsSimpleImpute$to_wide(columns = c("p.value", "FDR"))

Arguments

columns

value column default p.value

Returns

data.frame

Method clone()

The objects of this class are cloneable with this method.

Usage

ContrastsSimpleImpute$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.