Count distinct elements for each level of hierarchy per sample

Count distinct elements for each level of hierarchy per sample

Count distinct elements for each level of hierarchy per sample

Value

An R6 class generator.

Details

Provides wide, long, and plot views of hierarchy counts.

See also

Public fields

.summary

summarised data frame

.configuration

AnalysisConfiguration object

Methods

Public methods

HierarchyCountsSample$new()

HierarchyCountsSample$wide()

HierarchyCountsSample$long()

HierarchyCountsSample$plot()

HierarchyCountsSample$clone()

Method new()

Create a new HierarchyCountsSample

Usage

HierarchyCountsSample$new(pdata, configuration, nr_children = 1)

Arguments

pdata

data frame

configuration

AnalysisConfiguration

nr_children

minimum number of children to include

Method wide()

Return wide-format summary

Usage

HierarchyCountsSample$wide()

Method long()

Return long-format summary

Usage

HierarchyCountsSample$long()

Method plot()

Return barplot of hierarchy counts

Usage

HierarchyCountsSample$plot()

Method clone()

The objects of this class are cloneable with this method.

Usage

HierarchyCountsSample$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

bb <- sim_lfq_data_protein_config()
#> creating sampleName from file_name column
#> completing cases
#> completing cases done
#> setup done
counts <- HierarchyCountsSample$new(bb$data, bb$config)
head(counts$wide())
#> # A tibble: 6 × 3
#> # Groups:   isotopeLabel [1]
#>   isotopeLabel sampleName protein_Id
#>   <chr>        <chr>           <int>
#> 1 light        A_V1                9
#> 2 light        A_V2                9
#> 3 light        A_V3                7
#> 4 light        A_V4               10
#> 5 light        B_V1                9
#> 6 light        B_V2                9