Decorates LFQData with methods to compute statistics of interactions

Decorates LFQData with methods to compute statistics of interactions

Decorates LFQData with methods to compute statistics of interactions

Details

compute stdv, mean and CV per peptide or protein and condition.

See also

Public fields

lfq

LFQData

stat

either CV or sd (if is_transformed)

statsdf

frame with statistics.

Methods

Public methods

LFQDataStats$new()

LFQDataStats$stats()

LFQDataStats$stats_quantiles()

LFQDataStats$density()

LFQDataStats$density_median()

LFQDataStats$violin()

LFQDataStats$violin_median()

LFQDataStats$stdv_vs_mean()

LFQDataStats$power_t_test_quantiles()

LFQDataStats$power_t_test()

LFQDataStats$clone()

Method new()

create analyse variances and CV

Usage

LFQDataStats$new(lfqdata, stats = c("interaction", "all", "pooled"))

Arguments

lfqdata

LFQData object

Method stats()

compute CV sd and mean of e.g. peptide protein in each condition.

Usage

LFQDataStats$stats(all = FALSE)

Returns

data.frame

Method stats_quantiles()

Determine CV or sd for the quantiles

Usage

LFQDataStats$stats_quantiles(probs = c(0.1, 0.25, 0.5, 0.75, 0.9))

Arguments

probs

for which quantile to determine CV or sd

Method density()

plots density or ecdf

Usage

LFQDataStats$density(ggstat = c("density", "ecdf"))

Arguments

ggstat

either density or ecdf

Returns

ggplot

Method density_median()

plot density or ecdf of CV or sd for the 50

Usage

LFQDataStats$density_median(ggstat = c("density", "ecdf"))

Arguments

ggstat

either density of ecdf

Returns

ggplot

Method violin()

plot violinplot of CV or sd

Usage

LFQDataStats$violin()

Arguments

ggstat

either density of ecdf

Returns

ggplot

Method violin_median()

plot violinplot of CV or sd for the 50

Usage

LFQDataStats$violin_median()

Returns

ggplot

Method stdv_vs_mean()

plot sd vs mean

Usage

LFQDataStats$stdv_vs_mean(size = 200)

Arguments

size

number of points to sample (default 200)

Returns

ggplot

Method power_t_test_quantiles()

compute sample size for entire dataset

Usage

LFQDataStats$power_t_test_quantiles(
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9),
  delta = c(0.59, 1, 2),
  power = 0.8,
  sig.level = 0.05
)

Arguments

probs

quantiles of sd for which sample size should be computed

delta

effect size

power

power of test

sig.level

significance level.

Method power_t_test()

compute sample for each protein

Usage

LFQDataStats$power_t_test(delta = c(0.59, 1, 2), power = 0.8, sig.level = 0.05)

Arguments

delta

effect size

power

power of test

sig.level

significance level.

Method clone()

The objects of this class are cloneable with this method.

Usage

LFQDataStats$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# study variance of not normalized data
#source("c:/Users/wewol/prog/prolfqua/R/LFQData.R")
runallfuncs <- function(x){

  stopifnot("data.frame" %in% class(x$stats()))
  stopifnot(c("long", "wide") %in% names(x$stats_quantiles()))
  stopifnot("ggplot" %in% class(x$density()))
  stopifnot("ggplot" %in% class(x$density_median()))
  stopifnot("ggplot" %in% class(x$density("ecdf")))
  stopifnot("ggplot" %in% class(x$density_median("ecdf")))
  stopifnot("ggplot" %in% class(x$violin()))
  stopifnot("ggplot" %in% class(x$violin_median()))
  stopifnot("ggplot" %in% class(x$stdv_vs_mean(size = 400)))
  if(!x$lfq$is_transformed()){
    stopifnot(is.null(x$power_t_test()))
    stopifnot(is.null(x$power_t_test_quantiles()))
  }
}

istar <- prolfqua::data_ionstar$filtered()

#> Column added : nr_peptide_Id_IN_protein_Id

data <- istar$data %>% dplyr::filter(protein_Id %in% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
lfqstats <- lfqdata$get_Stats()

#> completing cases

runallfuncs(lfqstats)

#> Warning: `fun.y` is deprecated. Use `fun` instead.

#> Warning: data is not transformed - aborting

#> Warning: data is not transformed - aborting

x<-lfqstats

#study variance of normalized data
istar <- prolfqua::data_ionstar$normalized()
istar$config$table$is_intensity_transformed

#> [1] TRUE

data <- istar$data %>% dplyr::filter(protein_Id %in% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
lfqdata$is_transformed(TRUE)
lfqstats <- lfqdata$get_Stats()

#> completing cases

runallfuncs(lfqstats)

#> Warning: `fun.y` is deprecated. Use `fun` instead.

#Slightly different dataset
bb <- prolfqua::data_ionstar$filtered()

#> Column added : nr_peptide_Id_IN_protein_Id

stopifnot(nrow(bb$data) == 25780)
config <- bb$config$clone(deep = TRUE)
analysis <- bb$data

lfqdata <- LFQData$new(analysis, config)

lfqstats <- lfqdata$get_Stats()

#> completing cases

runallfuncs(lfqstats)

#> Warning: `fun.y` is deprecated. Use `fun` instead.

#> Warning: data is not transformed - aborting

#> Warning: data is not transformed - aborting