Write LFQData, or provide outputs for writing.

Write LFQData, or provide outputs for writing.

Write LFQData, or provide outputs for writing.

See also

Public fields

lfq

LFQData

format

format to write to

prefix

prefix of filename

file_paths

list with paths were data was written to.

Methods

Public methods

LFQDataWriter$new()

LFQDataWriter$get_long()

LFQDataWriter$get_wide()

LFQDataWriter$write_long()

LFQDataWriter$write_wide()

LFQDataWriter$clone()

Method new()

initialize class

Usage

LFQDataWriter$new(lfqdata, prefix = "ms_", format = "xlsx")

Arguments

lfqdata

LFQData

prefix

prefix files with

format

which format to write to ("xlsx", "csv", "html")

Method get_long()

Get data in long format for writing

Usage

LFQDataWriter$get_long()

Returns

tibble gets data formatted for writing

Method get_wide()

Get data in Wide format for writing

Usage

LFQDataWriter$get_wide()

Returns

list with data and annotation gets data formatted for writing

Method write_long()

write data to file

Usage

LFQDataWriter$write_long(path_qc)

Arguments

path_qc

path to write to

Method write_wide()

write data to file

Usage

LFQDataWriter$write_wide(path_qc)

Arguments

path_qc

path to write to

Method clone()

The objects of this class are cloneable with this method.

Usage

LFQDataWriter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.