Decorates LFQData with a row annotation and some protein specific functions.

Decorates LFQData with a row annotation and some protein specific functions.

Decorates LFQData with a row annotation and some protein specific functions.

See also

Public fields

row_annot

data.frame containing further information

pID

column with protein ids

description

name of column containing descriptions

ids

vector with columns containing addition IDs

nr_peptides

name of columns with the number of peptides

Methods

Public methods

ProteinAnnotation$new()

ProteinAnnotation$annotate_decoys()

ProteinAnnotation$annotate_contaminants()

ProteinAnnotation$nr_clean()

ProteinAnnotation$clean()

ProteinAnnotation$clone()

Method new()

initialize

Usage

ProteinAnnotation$new(
  lfqdata,
  row_annot,
  description = NULL,
  ids = NULL,
  nr_peptides = "nr_peptides"
)

Arguments

lfqdata

data frame from setup_analysis

row_annot

data frame with row annotation. Must have columns matching config$table$hierarchy_keys_depth()

description

name of column with description

ids

names of columns with additional ID's

nr_peptides

additional peptides

Method annotate_decoys()

annotate rev sequences

Usage

ProteinAnnotation$annotate_decoys(pattern = "REV_")

Arguments

pattern

default "REV_"

Method annotate_contaminants()

annotate contaminants

Usage

ProteinAnnotation$annotate_contaminants(pattern = "^zz|^CON")

Arguments

pattern

default "^zz|^CON"

Method nr_clean()

get number of neither contaminants nor decoys

Usage

ProteinAnnotation$nr_clean(contaminants = TRUE, decoys = TRUE)

Arguments

contaminants

remove contaminants

decoys

remove decoys return number of cleans

Method clean()

remove REV and CON sequences

Usage

ProteinAnnotation$clean(contaminants = TRUE, decoys = TRUE)

Arguments

contaminants

remove contaminants

decoys

remove decoys

Method clone()

The objects of this class are cloneable with this method.

Usage

ProteinAnnotation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

istar <- prolfqua_data('data_ionstar')$filtered()
#> Column added : nr_peptide_Id_IN_protein_Id
istar$config <- old2new(istar$config)
data <- istar$data #|> dplyr::filter(protein_Id %in% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
pannot <- ProteinAnnotation$new( lfqdata )

pannot$annotate_decoys()
#> [1] 0
pannot$annotate_contaminants()
#> [1] 5

stopifnot(pannot$nr_clean(contaminants = FALSE) == 163)
stopifnot(pannot$nr_clean(contaminants = TRUE) == 158)
stopifnot(nrow(pannot$clean(contaminants = FALSE)) == 163)
stopifnot(nrow(pannot$clean(contaminants = TRUE)) == 158)
dd <- pannot$clean()
tmp <- lfqdata$get_subset(dd)
#> Joining with `by = join_by(protein_Id)`
stopifnot(tmp$hierarchy_counts()$protein_Id == 158)