Decorates LFQData with a row annotation and some protein specific functions.

Decorates LFQData with a row annotation and some protein specific functions.

Decorates LFQData with a row annotation and some protein specific functions.

See also

Public fields

row_annot

data.frame containing further information

pID

column with protein ids

Methods

Public methods

RowAnnotProtein$new()

RowAnnotProtein$annotateREV()

RowAnnotProtein$annotateCON()

RowAnnotProtein$nr_clean()

RowAnnotProtein$clean()

RowAnnotProtein$clone()

Method new()

initialize

Usage

RowAnnotProtein$new(lfqdata, row_annot)

Arguments

lfqdata

data frame from setup_analysis

row_annot

data frame with row annotation. Must have columns matching config$table$hkeysDepth()

Method annotateREV()

annotate rev sequences

Usage

RowAnnotProtein$annotateREV(pattern = "REV_")

Arguments

pattern

default "REV_"

Method annotateCON()

annotate contaminants

Usage

RowAnnotProtein$annotateCON(pattern = "^zz|^CON")

Arguments

pattern

default "^zz|^CON"

Method nr_clean()

return number of cleans

Usage

RowAnnotProtein$nr_clean()

Method clean()

remove REV and CON sequences

Usage

RowAnnotProtein$clean()

Method clone()

The objects of this class are cloneable with this method.

Usage

RowAnnotProtein$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

istar <- prolfqua_data('data_ionstar')$filtered()
#> Column added : nr_peptide_Id_IN_protein_Id

data <- istar$data |> dplyr::filter(protein_Id %in% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
pannot <- RowAnnotProtein$new( lfqdata )

pannot$annotateREV()
#> [1] 0
pannot$annotateCON()
#> [1] 0

pannot$nr_clean()
#> [1] 62
dd <- pannot$clean()
tmp <- lfqdata$get_subset(dd)
#> Joining, by = "protein_Id"
tmp$complete_cases()
#> completing cases