Generate instances of AnalysisConfiguration

configurations examples of or various signal processing software outputs

file must be read with tidyMQ_Peptides, you will still need to add the factors (explanatory variables).

Usage

create_config_Skyline(
  isotopeLabel = "Isotope.Label",
  ident_qValue = "annotation_QValue"
)

create_config_Spectronaut_Peptide(
  isotopeLabel = "Isotope.Label",
  ident_qValue = "EG.Qvalue"
)

create_config_MQ_peptide(
  ident_qValue = "pep",
  intensity = "peptide.intensity",
  isotopeLabel = "isotope"
)

Arguments

isotopeLabel

isotope

ident_qValue

pep

intensity

peptide.intensity

See also

Examples

skylineconfig <- create_config_Skyline()
skylineconfig$factors[["Time"]] = "Sampling.Time.Point"
skylineconfig$factor_keys()
#> [1] "Time"
skylineconfig$hierarchy_keys()
#> [1] "protein_Id"   "peptide_Id"   "precursor_Id" "fragment_Id" 

spectronautconfig <- create_config_Spectronaut_Peptide()
config <- create_config_Spectronaut_Peptide()
config$factors[["coding"]] = "coding"
config$factors[["sex"]] = "sex"
config$factors[["age"]] = "age"
config$factors[["Sample_id"]] = "Sample.Name"

tmp <- create_config_MQ_peptide()