Compute theoretical sample sizes from factor level standard deviations

Compute theoretical sample sizes from factor level standard deviations

Usage

lfq_power_t_test_proteins(
  stats_res,
  delta = c(0.59, 1, 2),
  power = 0.8,
  sig.level = 0.05,
  min.n = 1.5
)

Arguments

stats_res

data.frame `summarize_stats` output

delta

effect size you are interested in

power

of test

sig.level

P-Value

min.n

smallest n to determine

See also

Examples

bb1 <- prolfqua::sim_lfq_data_peptide_config()
#> creating sampleName from file_name column
#> completing cases
#> completing cases done
#> setup done
ldata <- LFQData$new(bb1$data, bb1$config)
stats_res <- summarize_stats(ldata)
bb <- lfq_power_t_test_proteins(stats_res)