Compute theoretical sample sizes from factor level standard deviations

Compute theoretical sample sizes from factor level standard deviations

Usage

lfq_power_t_test_quantiles(
  pdata,
  config,
  delta = 1,
  power = 0.8,
  sig.level = 0.05,
  probs = seq(0.5, 0.9, by = 0.1)
)

Arguments

pdata

data.frame

config

AnalysisConfiguration

delta

effect size you are interested in

power

of test

sig.level

P-Value

probs

numeric vector of quantile probabilities

See also

Examples

bb1 <- prolfqua::sim_lfq_data_peptide_config()
#> creating sampleName from fileName column
#> completing cases
#> completing cases done
#> setup done
config <- bb1$config
data2 <- bb1$data

res <- lfq_power_t_test_quantiles(data2, config)
#> Warning: Intensities are not transformed yet.
#> completing cases
res$summary
#> # A tibble: 5 × 4
#>   quantile probs    sd `FC=2`
#>   <chr>    <dbl> <dbl>  <dbl>
#> 1 50%        0.5 0.861     13
#> 2 60%        0.6 1.00      17
#> 3 70%        0.7 1.08      20
#> 4 80%        0.8 1.28      27
#> 5 90%        0.9 1.48      36
stats_res <- summarize_stats(data2, config)
#> completing cases
res <- lfq_power_t_test_quantiles(data2, config, delta = 2)
#> Warning: Intensities are not transformed yet.
#> completing cases
res <- lfq_power_t_test_quantiles(data2, config, delta = c(0.5,1,2))
#> Warning: Intensities are not transformed yet.
#> completing cases