Marks peptides which do not correlate with other peptides of a protein

It computes the pairwise correlation among all peptides and marks those with with a corralation less then minCorrelation to all other peptides

mark_decorelated(data, config, minCorrelation = 0.7)

Arguments

data

data

config

configuration

minCorrelation

smallest allowed correlation

TODO

do something with warnings of type "the standard deviation is zero".

do investigate In max(x, na.rm = TRUE) : no non-missing arguments to max; returning -Inf

Examples

bb <- prolfqua_data('data_ionstar')$filtered()
#> Column added : nr_peptide_Id_IN_protein_Id
stopifnot( nrow(bb$data) == 25780)
conf <- old2new(bb$config$clone(deep=TRUE))
data <- bb$data |> dplyr::ungroup()
dim(data)
#> [1] 25780    10
dataI <- mark_decorelated(data, conf)
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 7: `row.names = "ASVHTLSGHTNAVATVR"` and `col.names =
#>   "VYREDDTATEETHPVSWKPEIIK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "DVLAHQVPNAK"` and `col.names =
#>   "LHTKPSQAPAVEVAPAGASYNPSFEDHQTLLSAAHEVELQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 9 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "AQLSQYFDAVK"` and `col.names = "FVPTALR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 8 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 75: `row.names = "NCPHVVVGTPGR"` and `col.names =
#>   "VNIVFNYDMPEDSDTYLHR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 60: `row.names = "IPELDMTEVVAPFMANIPLLLYPQDGPR"` and `col.names =
#>   "TEKEILDAFDK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 10 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 31: `row.names = "FEAAETLEEAAMR"` and `col.names = "YEEVEYYYQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 9 remaining warnings.
#> Warning: There were 14 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 5: `row.names = "AIPNQGEILVIR"` and `col.names =
#>   "IEGSGDQVDTLELSGGAR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 13 remaining warnings.
#> Warning: There were 268 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 8: `row.names = "AFQQNLVSSALFSK"` and `col.names =
#>   "HHVLQNDVLAHQSTVEAVNK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 267 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 30: `row.names = "GDEVIFGEFSWPK"` and `col.names = "SAPLEIGLQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 33 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "DFNPTATVK"` and `col.names = "EGLLFEGR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 32 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 94: `row.names = "KAEAQIAAK"` and `col.names =
#>   "MNLGVGAYRDDNGKPYVLPSVR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 67 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 29: `row.names = "AEFVEVTK"` and `col.names = "VHKECCHGDLLECADDR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 66 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "DQQNLPYGVTPASPSGHSQGR"` and `col.names =
#>   "TTPTPSTNSVLSTSTNR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "LLDGFFIRPFYK"` and `col.names =
#>   "RPTPQDNLTDAENGNIQLQAQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 8: `row.names = "GVVVQDAALLESGAAIR"` and `col.names =
#>   "TAEGGAEHVQPITGDNIVNVLDDFR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 19 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 9: `row.names = "ANKLDHVVTIIK"` and `col.names = "VEDLTFTSPFCLQVK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 18 remaining warnings.
#> Warning: There were 5 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "FALEVPGLR"` and `col.names = "LSETSPAFR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 4 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "MTEWETAAPAVAETPDIK"` and `col.names =
#>   "QAVDVSPLRR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 33 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "AIEQLKEQAATGK"` and `col.names = "ATIFNLK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 32 remaining warnings.
#> Warning: There were 6 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AAIQSQLDGVR"` and `col.names = "KYVLDDLIVAK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 5 remaining warnings.
#> Warning: There were 11 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "ELQHEKDTSSLLR"` and `col.names = "ICSLQEEQQQLR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 10 remaining warnings.
#> Warning: There were 8 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "AGQMVTVWAAGAGVAHSPPSTLVWK"` and `col.names =
#>   "ELEEAMAGERDK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 7 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AELDLMAPGPLPR"` and `col.names =
#>   "VTNTTPWEHPYPDLPSR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 387: `row.names = "GPLEYVPSAK"` and `col.names = "TAVFGFETSEAK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "EVAWEQNYPDR"` and `col.names = "ITGIDSSPAMIAEAR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 6 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 220: `row.names = "VLPLEEAYR"` and `col.names =
#>   "VYIGSFWAQPLQNTDNRR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 5 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 121: `row.names = "HIAEDSDRKYEEVAR"` and `col.names = "MELQEMQLK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 7: `row.names = "AQEAPGQAEPPAAAEVQGAGNENEPR"` and `col.names =
#>   "LLVVDRETDEFFK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AQYVLAEQVTR"` and `col.names =
#>   "FQQAGHKPVALVGGATGLIGDPSFK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "ASAGSYISALR"` and `col.names = "GPGLYYVDEHGTR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 146: `row.names = "ETAPGDISPYTIPEEDR"` and `col.names =
#>   "PVMQALLHIIR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "LKTEEGEIDYSAEEGENRR"` and `col.names =
#>   "TIPGTALVEMGDEYAVER"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 6 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 27: `row.names = "FSLFAERK"` and `col.names =
#>   "QLSCVANQNGSQADCELGNPFKR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 5 remaining warnings.
#> Warning: There were 10 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AFAASNELEVNATHKDR"` and `col.names =
#>   "EGITATDLVLTVTQMLR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 9 remaining warnings.
#> Warning: There were 7 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 1: `row.names = "AIQAELLK"` and `col.names = "AIQAELLK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 6 remaining warnings.
#> Warning: There were 16 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 52: `row.names = "EEMLQMIR"` and `col.names = "KANYAVDAYFR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 15 remaining warnings.
#> Warning: There were 14 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 577: `row.names = "DHRDNLEFFLAGIGR"` and `col.names =
#>   "FCFTPHTEEGCLSER"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 13 remaining warnings.
#> Warning: There were 3 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 1: `row.names = "FSIALLNQAVER"` and `col.names = "FSIALLNQAVER"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 2 remaining warnings.
#> Warning: There were 8 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 5: `row.names = "DFYLATSPPDSFLDDHHLTRPHPER"` and `col.names =
#>   "NHQDPLAVAYHLIIDNR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 7 remaining warnings.
#> Warning: There were 3 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "FIDTSQFILNR"` and `col.names = "SKPVFSESLSD"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 2 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 335: `row.names = "EKLCYVALDFEQEMATAASSSSLEK"` and `col.names =
#>   "HQGVMVGMGQKDSYVGDEAQSK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 20: `row.names = "GPRPPNQQQPIEGTDRVEPK"` and `col.names =
#>   "SQADKPVLAIQVLGTVK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 6: `row.names = "ITYVANVNPGGWAPASVLR"` and `col.names =
#>   "RHGSMVSLVSGASGYSATSTSSFK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "GLSNLFLSCPIPK"` and `col.names = "LLLLAGVDR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 18 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 37: `row.names = "ALSINTLVK"` and `col.names = "FASTVFPSDHIPSR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 17 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 8: `row.names = "ESTSYIEEHIDR"` and `col.names =
#>   "TQYLQSYGPEHLLTFSNLR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: `cols` is now required when using `unnest()`.
#> ℹ Please use `cols = c(srmDecor)`.