compute per group averages and impute values should generalize at some stage

compute per group averages and impute values should generalize at some stage

missigness_impute_factors_interactions(
  pdata,
  config,
  probs = 0.03,
  value = c("long", "nrReplicates", "nrMeasured", "meanAbundance", "imputed"),
  add.prefix = FALSE,
  global = TRUE
)

Arguments

pdata

data.frame

config

AnalysisConfiguration

probs

quantile to take average from (default 0.1)

value

use default

add.prefix

use default

See also

Examples

istar <- sim_lfq_data_peptide_config()
#> sample already exists
#> completing cases
config <- istar$config
analysis <- istar$data

xx <- complete_cases(analysis, config)
#> completing cases

res <- missigness_impute_factors_interactions(xx, config)
#> completing cases
res <- missigness_impute_factors_interactions(xx, config, value = "imputed")
#> completing cases
res <- missigness_impute_factors_interactions(xx, config, value = "nrMeasured")
#> completing cases
long <- missigness_impute_factors_interactions(xx, config, value = "long")
#> completing cases