pivot model contrasts matrix to wide format produced by `contrasts_linfct` and ...

pivot model contrasts matrix to wide format produced by `contrasts_linfct` and ...

pivot_model_contrasts_2_Wide(
  modelWithInteractionsContrasts,
  subject_Id = "protein_Id",
  columns = c("estimate", "p.value", "p.value.adjusted"),
  contrast = "lhs"
)

See also

Examples

dd <- prolfqua::data_factor_levelContrasts
head(dd)

#> # A tibble: 6 x 14
#>   Compound    lhs      sigma    df estimate std.error statistic p.value conf.low
#>   <chr>       <chr>    <dbl> <int>    <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#> 1 4.85_649.0~ Mortali~ 0.987   115   0.128      0.193     0.659   0.511   -0.256
#> 2 4.85_649.0~ Interve~ 0.987   115   0.0384     0.183     0.210   0.834   -0.324
#> 3 4.85_649.0~ NRSNRS_~ 0.987   115   0.173      0.222     0.777   0.439   -0.268
#> 4 4.85_649.0~ NRSNRS_~ 0.987   115  -0.0471     0.222    -0.212   0.833   -0.487
#> 5 4.85_649.0~ NRSNRS_~ 0.987   115  -0.220      0.221    -0.993   0.323   -0.658
#> 6 5.05_635.0~ Mortali~ 0.719   115   0.104      0.141     0.741   0.460   -0.175
#> # ... with 5 more variables: conf.high <dbl>, isSingular <lgl>,
#> #   sigma.model <dbl>, df.residual.model <dbl>, p.value.adjusted <dbl>

tmp <- pivot_model_contrasts_2_Wide(dd, subject_Id = "Compound")
tmp

#> # A tibble: 30 x 16
#>    Compound  `estimate.Inter~ `estimate.Morta~ `estimate.NRSNR~ `estimate.NRSNR~
#>    <chr>                <dbl>            <dbl>            <dbl>            <dbl>
#>  1 4.85_649~          0.0384            0.128            0.173           -0.0471
#>  2 5.05_635~          0.00972           0.104            0.118           -0.0229
#>  3 5.11_366~          0.245            -0.278            0.367            0.205 
#>  4 5.12_260~          0.160            -0.431            0.0772           0.0195
#>  5 5.12_302~          0.370            -0.253            0.303            0.223 
#>  6 5.12_476~         -0.00471           0.436           -0.0256           0.436 
#>  7 5.13_363~         -0.626            -0.255           -0.403            0.264 
#>  8 5.16_166~          0.236            -0.283           -0.346           -0.154 
#>  9 5.16_242~          0.176            -0.0713          -0.270            0.175 
#> 10 5.16_277~         -0.0342           -0.132            0.0999           0.338 
#> # ... with 20 more rows, and 11 more variables:
#> #   estimate.NRSNRS_4 - NRSNRS_5 <dbl>,
#> #   p.value.InterventionCTRL - InterventionINT <dbl>,
#> #   p.value.MortalityM0 - MortalityM1 <dbl>, p.value.NRSNRS_3 - NRSNRS_4 <dbl>,
#> #   p.value.NRSNRS_3 - NRSNRS_5 <dbl>, p.value.NRSNRS_4 - NRSNRS_5 <dbl>,
#> #   p.value.adjusted.InterventionCTRL - InterventionINT <dbl>,
#> #   p.value.adjusted.MortalityM0 - MortalityM1 <dbl>,
#> #   p.value.adjusted.NRSNRS_3 - NRSNRS_4 <dbl>,
#> #   p.value.adjusted.NRSNRS_3 - NRSNRS_5 <dbl>,
#> #   p.value.adjusted.NRSNRS_4 - NRSNRS_5 <dbl>