Add protein lengths from fasta file to data frame (id_col - protein id column.)

Add protein lengths from fasta file to data frame (id_col - protein id column.)

Convert tidy table with protein quants into SaintExpress compatible inputs

Network visualization. look at https://www.jessesadler.com/post/network-analysis-with-r/ https://fgcz-intranet.uzh.ch/tiki-index.php?page=WG_APMSnProximityLabeling

add_protein_lengths(intdata, fasta, id_col = "protein_Id")

protein_2localSaint(
  xx,
  quantcolumn = "mq.protein.intensity",
  proteinID = "protein_Id",
  geneNames = proteinID,
  proteinLength = "protein.length",
  IP_name = "raw.file",
  baitCol = "bait",
  CorTCol = "CorT"
)

runSaint(si, filedir = getwd(), spc = TRUE, CLEANUP = TRUE)

Arguments

intdata

data.frame

fasta

list of sequences created with read.fasta

id_col

column with protein ids/accessions.

xx

data.frame in long format

quantcolumn

intensity column

proteinID

protein accession

geneNames

column with gene names

proteinLength

column with protein lengths

IP_name

raw.file

baitCol

column with bait definition (condition)

CorTCol

is it control or TRUE (SaintExpress speach)

si

output of protein_2localSaint function

filedir

where to store the saint express inputs

spc

if TRUE spectum counts if FALSE intensities (see SaintExpress documentation for more details)

CLEANUP

TRUE remove all files generated bySaintExpress.

Examples

bb <- prolfqua_data('data_IonstarProtein_subsetNorm')
bb$config <- old2new(config = bb$config$clone( deep = TRUE))
xx <- LFQData$new(bb$data, bb$config)
exampleDat <- xx$data |> dplyr::mutate(CorT = dplyr::case_when(dilution. == "a" ~ "C", TRUE ~ "T"))
# sample protein lengths

tmp <- data.frame(protein_Id = unique(exampleDat$protein_Id))
tmp$proteinLength <- as.integer(runif(nrow(tmp), min = 150, max = 2500))
exampleDat <- dplyr::inner_join(tmp, exampleDat)
#> Joining, by = "protein_Id"
#undebug(protein_2localSaint)
res <- protein_2localSaint(exampleDat,quantcolumn = "medpolish",
                   proteinID = "protein_Id",
                   proteinLength = "proteinLength",
                   IP_name = "raw.file",
                   baitCol = "dilution.",
                   CorTCol = "CorT"
                   )

stopifnot(names(res) == c( "inter", "prey",  "bait"))
if(!Sys.info()["sysname"] == "Darwin") {
  data_SAINTe_output <- runSaint(res, filedir = tempdir())
  #usethis::use_data(data_SAINTe_output)
} else {
 testthat::expect_error(runSaint(res, filedir = tempdir()))
}
#> /var/folders/5v/30q3lqxs6_b88gc005sg2t7c0000gn/T//Rtmpy1ri86/inter.txt
#> /var/folders/5v/30q3lqxs6_b88gc005sg2t7c0000gn/T//Rtmpy1ri86/prey.txt
#> /var/folders/5v/30q3lqxs6_b88gc005sg2t7c0000gn/T//Rtmpy1ri86/bait.txt