Spreads isotope label heavy and light into two columns

Spreads isotope label heavy and light into two columns

Usage

spread_response_by_IsotopeLabel(resData, config)

See also

Examples

bb <- prolfqua::sim_lfq_data_peptide_config()
#> creating sampleName from fileName column
#> completing cases
#> completing cases done
#> setup done
configur <- bb$config$clone(deep=TRUE)
data <- bb$data

x<-spread_response_by_IsotopeLabel(data,configur)

bb <- prolfqua_data('data_skylineSRM_HL_A')
configur <- bb$config_f()
#> config$table is deprecated, use config directly
#> config$table is deprecated, use config directly
#> config$table is deprecated, use config directly
data <- bb$analysis(bb$data, configur)
#> creating sampleName from fileName column
#> Warning: no nr_children column specified in the data, adding column nr_children and setting to 1.
#> completing cases
#> completing cases done
#> setup done

bb <- prolfqua_data('data_skylineSRM_HL_A')
conf <- bb$config_f()
#> config$table is deprecated, use config directly
#> config$table is deprecated, use config directly
#> config$table is deprecated, use config directly
analysis <- bb$analysis(bb$data, bb$config_f())
#> config$table is deprecated, use config directly
#> config$table is deprecated, use config directly
#> config$table is deprecated, use config directly
#> creating sampleName from fileName column
#> Warning: no nr_children column specified in the data, adding column nr_children and setting to 1.
#> completing cases
#> completing cases done
#> setup done
x <- spread_response_by_IsotopeLabel(analysis, conf)