Create Annotation

Annotates Data Table

See also

Public fields

fileName

column name of column containing raw file names

sampleName

(will be generated from factors or fileName)

isotopeLabel

which column contains the isotope label (e.g. heavy or light), Gor light only if LFQ.

ident_qValue

column name with identification QValues (smaller better)

ident_Score

column with identification score (lager better)

opt_rt

optional column with rt information

opt_mz

optional column with mz information

nr_children

optional column containing for instance the number of peptides

workIntensity

column which contains the intensities

is_response_transformed

are the intensities transformed for constant variance

factors

Names of columns containing factors (annotations)

factorDepth

number of relevant factors (used by plotting functions etc)

hierarchy

list with columns describing the measurement hierarchy (i.e. protein peptide precursor fragment)

hierarchyDepth

At which depth do you want to model i.e. i.e. protein than 1

Methods

Public methods

AnalysisTableAnnotation$new()

AnalysisTableAnnotation$set_response()

AnalysisTableAnnotation$get_response()

AnalysisTableAnnotation$pop_response()

AnalysisTableAnnotation$factor_keys()

AnalysisTableAnnotation$factor_keys_depth()

AnalysisTableAnnotation$hierarchy_keys()

AnalysisTableAnnotation$hierarchyKeys()

AnalysisTableAnnotation$hierarchy_keys_depth()

AnalysisTableAnnotation$hkeysDepth()

AnalysisTableAnnotation$id_required()

AnalysisTableAnnotation$id_vars()

AnalysisTableAnnotation$value_vars()

AnalysisTableAnnotation$annotation_vars()

AnalysisTableAnnotation$clone()

Method new()

create a new AnalysisTableAnnotationG

Usage

AnalysisTableAnnotation$new()

Method set_response()

Add name of intensity column

Usage

AnalysisTableAnnotation$set_response(colName)

Arguments

colName

name of intensity column

Method get_response()

Get name of working intensity column

Usage

AnalysisTableAnnotation$get_response()

Method pop_response()

Remove last name in array of working intensity column names

Usage

AnalysisTableAnnotation$pop_response()

Method factor_keys()

Get factor keys

Usage

AnalysisTableAnnotation$factor_keys()

Returns

array with keys

Method factor_keys_depth()

Get factor keys till factorDepth

Usage

AnalysisTableAnnotation$factor_keys_depth()

Method hierarchy_keys()

get hierarchy keys

Usage

AnalysisTableAnnotation$hierarchy_keys(rev = FALSE)

Arguments

rev

return in reverse order

Returns

array of column names

Method hierarchyKeys()

get hierarchy keys

Usage

AnalysisTableAnnotation$hierarchyKeys(rev = FALSE)

Arguments

rev

return in reverse order

Returns

array of column names

Method hierarchy_keys_depth()

get hierarchy keys up to depth

Usage

AnalysisTableAnnotation$hierarchy_keys_depth(names = TRUE)

Arguments

names

if TRUE names only if FALSE key value pairs

Returns

array of column names

Method hkeysDepth()

get hierarchy keys up to depth

Usage

AnalysisTableAnnotation$hkeysDepth(names = TRUE)

Arguments

names

if TRUE names only if FALSE key value pairs

Returns

array of column names

Method id_required()

Id Columns which must be in the input data frame

Usage

AnalysisTableAnnotation$id_required()

Returns

character array

Method id_vars()

get names of columns annotating values (e.g. intensities)

Usage

AnalysisTableAnnotation$id_vars()

Returns

character array

Method value_vars()

get names of columns containing observations e.g. (intensity, qValue, mz or rt)

Usage

AnalysisTableAnnotation$value_vars()

Method annotation_vars()

get names of columns with sample annotations

Usage

AnalysisTableAnnotation$annotation_vars()

Method clone()

The objects of this class are cloneable with this method.

Usage

AnalysisTableAnnotation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

ata <- AnalysisTableAnnotation$new()
ata$fileName = "rawfile.column"
ata$hierarchy[["protein"]] = "protein.column"
ata$factors[["explanatory"]] = "explanatory.column"
ata$set_response("abundance")
ata$id_required()
#>                               explanatory              protein 
#>     "rawfile.column" "explanatory.column"     "protein.column" 
#>                      
#>       "isotopeLabel" 
ata$id_vars()
#> [1] "rawfile.column" "explanatory"    "protein"        "isotopeLabel"  
#> [5] "sampleName"    
ata$value_vars()
#> [1] "abundance"   "qValue"      "nr_children"
ata$annotation_vars()
#> [1] "rawfile.column" "sampleName"     "explanatory"   
ac <- AnalysisConfiguration$new(ata)