Create Annotation

Annotates Data Table

See also

Public fields

fileName

column name of column containing raw file names

sampleName

(will be generated from factors)

isotopeLabel

which column contains the isotope label (e.g. heavy or light), Gor light only if LFQ.

ident_qValue

column name with identification QValues (smaller better)

ident_Score

column with identification score (lager better)

opt_rt

optional column with rt information

opt_mz

optional column with mz information

workIntensity

column which contains the intensities

is_intensity_transformed

are the intensities transformed for constant variance

factors

names of columns containing factors (annotions)

factorDepth

facet plot according to the first or the first two factors (factorDepth can be 1 or 2)

hierarchy

list with columns describing the measurement hierarchy (i.e. protein peptide precursor fragment)

hierarchyDepth

At which depth do you want to model i.e. i.e. protein than 1

Methods

Public methods

AnalysisTableAnnotation$new()

AnalysisTableAnnotation$setWorkIntensity()

AnalysisTableAnnotation$getWorkIntensity()

AnalysisTableAnnotation$popWorkIntensity()

AnalysisTableAnnotation$factorKeys()

AnalysisTableAnnotation$fkeysDepth()

AnalysisTableAnnotation$hierarchyKeys()

AnalysisTableAnnotation$hkeysDepth()

AnalysisTableAnnotation$idRequired()

AnalysisTableAnnotation$idVars()

AnalysisTableAnnotation$valueVars()

AnalysisTableAnnotation$annotationVars()

AnalysisTableAnnotation$clone()

Method new()

create a new AnalysisTableAnnotationG

Usage

AnalysisTableAnnotation$new()

Method setWorkIntensity()

add name of intensity column

Usage

AnalysisTableAnnotation$setWorkIntensity(colName)

Arguments

colName

name of intensity column

Method getWorkIntensity()

get name of working intensity column

Usage

AnalysisTableAnnotation$getWorkIntensity()

Method popWorkIntensity()

remove last name in array of working intensity column names

Usage

AnalysisTableAnnotation$popWorkIntensity()

Method factorKeys()

get factor keys

Usage

AnalysisTableAnnotation$factorKeys()

Returns

array with keys

Method fkeysDepth()

get factor keys till factorDepth

Usage

AnalysisTableAnnotation$fkeysDepth()

Method hierarchyKeys()

get hierarchy keys

Usage

AnalysisTableAnnotation$hierarchyKeys(rev = FALSE)

Arguments

rev

return in reverse order

Returns

array of column names

Method hkeysDepth()

get hierarchy keys up to depth

Usage

AnalysisTableAnnotation$hkeysDepth(names = TRUE)

Arguments

names

if TRUE names only if FALSE key value pairs

Returns

array of column names

Method idRequired()

Id Columns which must be in the input data frame

Usage

AnalysisTableAnnotation$idRequired()

Returns

character array

Method idVars()

get names of columns annotating values (e.g. intensities)

Usage

AnalysisTableAnnotation$idVars()

Returns

character array

Method valueVars()

get names of columns containing observations e.g. (intensity, qValue, mz or rt)

Usage

AnalysisTableAnnotation$valueVars()

Method annotationVars()

get names of columns with sample annotations

Usage

AnalysisTableAnnotation$annotationVars()

Method clone()

The objects of this class are cloneable with this method.

Usage

AnalysisTableAnnotation$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.