Estimate contrasts using Wald Test

Estimate contrasts using Wald Test

Estimate contrasts using Wald Test

See also

Super class

prolfqua::ContrastsInterface -> Contrast

Public fields

models

Model

contrasts

character with contrasts

contrastfun

function to compute contrasts

modelName

model name

subject_Id

name of column containing e.g., protein Id's

p.adjust

function to adjust p-values (default prolfqua::adjust_p_values)

contrast_result

data frame containing results of contrast computation

global

use a global linear function (determined by get_linfct)

Methods

Public methods

Contrasts$new()

Contrasts$get_contrast_sides()

Contrasts$get_linfct()

Contrasts$get_contrasts()

Contrasts$get_Plotter()

Contrasts$to_wide()

Contrasts$clone()

Inherited methods

prolfqua::ContrastsInterface$column_description()

Method new()

initialize create Contrast

Usage

Contrasts$new(
  model,
  contrasts,
  p.adjust = prolfqua::adjust_p_values,
  global = TRUE,
  modelName = "WaldTest"
)

Arguments

model

a dataframe with a structure similar to that generated by build_model

contrasts

a character vector with contrast specificiation

p.adjust

function to adjust the p-values

global

development/internal argument (if FALSE determine linfct for each model.)

modelName

name of contrast method, default WaldTest

Method get_contrast_sides()

get both sides of contrasts

Usage

Contrasts$get_contrast_sides()

Method get_linfct()

get linear functions from contrasts

Usage

Contrasts$get_linfct(global = TRUE)

Arguments

global

logical TRUE - get the a linear functions for all models, FALSE - linear function for each model

Method get_contrasts()

get table with contrast estimates

Usage

Contrasts$get_contrasts(all = FALSE)

Arguments

all

should all columns be returned (default FALSE)

Returns

data.frame with contrasts

Method get_Plotter()

return ContrastsPlotter creates Contrast_Plotter

Usage

Contrasts$get_Plotter(FCthreshold = 1, FDRthreshold = 0.1)

Arguments

FCthreshold

fold change threshold to show in plots

FDRthreshold

FDR threshold to show in plots

Returns

ContrastsPlotter

Method to_wide()

convert to wide format

Usage

Contrasts$to_wide(columns = c("p.value", "FDR"))

Arguments

columns

value column default p.value

Returns

data.frame

Method clone()

The objects of this class are cloneable with this method.

Usage

Contrasts$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Fitting mixed effects model to peptide data
istar <- prolfqua_data('data_ionstar')$normalized()
istar$config <- old2new(istar$config)
istar_data <- dplyr::filter(istar$data ,protein_Id %in% sample(protein_Id, 20))

modelFunction <-
strategy_lmer("transformedIntensity  ~ dilution. + (1 | peptide_Id) + (1 | sampleName)")

config <- istar$config
config$table$hkeysDepth()
#> [1] "protein_Id"

mod <- build_model(
 istar$data,
 modelFunction,
 subject_Id = config$table$hkeysDepth())
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> Warning: WARN :Error: number of levels of each grouping factor must be < number of observations (problems: sampleName)
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> boundary (singular) fit: see help('isSingular')
#> Joining, by = "protein_Id"

prolfqua::model_summary(mod)
#> $exists
#> 
#> FALSE  TRUE 
#>     1   162 
#> 
#> $isSingular
#> 
#> FALSE  TRUE 
#>    79    83 
#> 
 Contr <- c("dil.a_vs_b" = "dilution.a - dilution.b",
    "dil.e_vs_b" = "dilution.e - dilution.b" )
#Contrasts$debug("get_contrasts")
contrastX <- prolfqua::Contrasts$new(mod, Contr)
contrastX$get_contrasts()
#> determine linear functions:
#> compute contrasts:
#> computing contrasts.
#> Warning: Model is rank deficient!
#> Warning: Model is rank deficient!
#> Joining, by = c("protein_Id", "contrast")
#> # A tibble: 320 × 13
#> # Groups:   contrast [2]
#>    model…¹ prote…² contr…³    diff     FDR std.e…⁴    df stati…⁵ conf.…⁶ conf.…⁷
#>    <chr>   <chr>   <chr>     <dbl>   <dbl>   <dbl> <dbl>   <dbl>   <dbl>   <dbl>
#>  1 WaldTe… sp|P21… dil.a_… -0.0975 5.74e-1  0.0559 263.   -1.74  -0.208   0.0127
#>  2 WaldTe… sp|P21… dil.e_…  0.0291 8.49e-1  0.0561 263.    0.518 -0.0814  0.140 
#>  3 WaldTe… sp|Q9N… dil.a_…  0.305  5.91e-1  0.164   11.4   1.86  -0.0540  0.663 
#>  4 WaldTe… sp|Q9N… dil.e_…  0.336  3.10e-1  0.178   11.6   1.89  -0.0520  0.725 
#>  5 WaldTe… sp|O43… dil.a_… -0.0245 9.86e-1  0.127   12.3  -0.193 -0.300   0.252 
#>  6 WaldTe… sp|O43… dil.e_… -0.0228 9.09e-1  0.136   16.7  -0.167 -0.311   0.265 
#>  7 WaldTe… sp|P0A… dil.a_… -0.926  4.41e-2  0.274   18.0  -3.38  -1.50   -0.351 
#>  8 WaldTe… sp|P0A… dil.e_…  0.846  4.40e-3  0.204   18.0   4.16   0.418   1.27  
#>  9 WaldTe… sp|P0A… dil.a_… -0.997  1.01e-3  0.200   27.0  -4.99  -1.41   -0.587 
#> 10 WaldTe… sp|P0A… dil.e_…  0.790  2.94e-4  0.156   27.0   5.06   0.470   1.11  
#> # … with 310 more rows, 3 more variables: p.value <dbl>, sigma <dbl>,
#> #   avgAbd <dbl>, and abbreviated variable names ¹​modelName, ²​protein_Id,
#> #   ³​contrast, ⁴​std.error, ⁵​statistic, ⁶​conf.low, ⁷​conf.high
contrastX$get_contrast_sides()
#> # A tibble: 2 × 3
#>   contrast   group_1    group_2   
#>   <chr>      <chr>      <chr>     
#> 1 dil.a_vs_b dilution.a dilution.b
#> 2 dil.e_vs_b dilution.e dilution.b
contrastX$column_description()
#>           column_name
#> modelName   modelName
#> contrast     contrast
#> avgAbd         avgAbd
#> diff             diff
#> FDR               FDR
#> statistic   statistic
#> std.error   std.error
#> df                 df
#> p.value       p.value
#> conf.low     conf.low
#> conf.high   conf.high
#> sigma           sigma
#>                                                                                            description
#> modelName                                                                                type of model
#> contrast                                                      name of difference e.g. group1_vs_group2
#> avgAbd                                                  mean abundance value of protein in all samples
#> diff                                                                       difference among conditions
#> FDR                                                                               false discovery rate
#> statistic                                                                                 t-statistics
#> std.error                                                                               standard error
#> df                                                                                  degrees of freedom
#> p.value                                                                                        p-value
#> conf.low                                                         lower value of 95 confidence interval
#> conf.high                                                         high value of 95 confidence interval
#> sigma     residual standard deviation of linear model (needed for empirical Bayes variance shrinkage).
stopifnot(dim(contrastX$to_wide()) == c(160,7))