plot contrasts

plot contrasts

plot contrasts

See also

Public fields

contrastDF

data frame with contrasts

modelName

of column with model name

subject_Id

hierarchy key columns

prefix

default Contrasts - used to generate file names

diff

column with fold change differences

contrast

column with contrasts names, default "contrast"

volcano_spec

volcano plot specification

score_spec

score plot specification

histogram_spec

plot specification

fcthresh

fold change threshold

avg.abundance

name of column containing avg abundance values.

Methods

Public methods

ContrastsPlotter$new()

ContrastsPlotter$histogram()

ContrastsPlotter$histogram_estimate()

ContrastsPlotter$histogram_diff()

ContrastsPlotter$volcano()

ContrastsPlotter$volcano_plotly()

ContrastsPlotter$ma_plot()

ContrastsPlotter$ma_plotly()

ContrastsPlotter$score_plot()

ContrastsPlotter$score_plotly()

ContrastsPlotter$barplot_threshold()

ContrastsPlotter$clone()

Method new()

create Crontrast_Plotter

Usage

ContrastsPlotter$new(
  contrastDF,
  subject_Id,
  volcano = list(list(score = "FDR", thresh = 0.1)),
  histogram = list(list(score = "p.value", xlim = c(0, 1, 0.05)), list(score = "FDR",
    xlim = c(0, 1, 0.05))),
  score = list(list(score = "statistic", thresh = NULL)),
  fcthresh = 1,
  modelName = "modelName",
  diff = "diff",
  contrast = "contrast",
  avg.abundance = "avgAbd"
)

Arguments

contrastDF

frame with contrast data

subject_Id

columns containing subject Identifier

volcano

which score to plot and which ablines to add.

histogram

which scores to plot and which range (x) should be shown.

score

score parameters

fcthresh

default 1 (log2 FC threshold)

modelName

name of column with model names

diff

fold change (difference) diff column

contrast

contrast column

avg.abundance

name of column with average abundance

Method histogram()

plot histogram of selected scores (e.g. p-value, FDR, t-statistics)

Usage

ContrastsPlotter$histogram()

Method histogram_estimate()

plot histogram of effect size - difference between groups

Usage

ContrastsPlotter$histogram_estimate(binwidth = 0.05)

Arguments

binwidth

with of bin in histogram

Method histogram_diff()

plot histogram of differences (diff) fold change

Usage

ContrastsPlotter$histogram_diff(binwidth = 0.05)

Arguments

binwidth

with of bin in histogram

Method volcano()

volcano plots (fold change vs FDR)

Usage

ContrastsPlotter$volcano(
  colour,
  legend = TRUE,
  scales = c("fixed", "free", "free_x", "free_y")
)

Arguments

colour

column name with color information default modelName

legend

default TRUE

scales

default fixed facet_wrap, scales argument

Method volcano_plotly()

plotly volcano plots

Usage

ContrastsPlotter$volcano_plotly(
  colour,
  legend = TRUE,
  scales = c("fixed", "free", "free_x", "free_y")
)

Arguments

colour

column in contrast matrix with colour coding

legend

default TRUE

scales

default fixed facet_wrap, scales argument

Returns

list of ggplots

Method ma_plot()

ma plot

MA plot displays the effect size estimate as a function of the mean protein intensity across groups. Each dot represents an observed protein. Red horizontal lines represent the fold-change threshold.

Sometimes measured effects sizes (differences between samples groups) are biased by the signal intensity (here protein abundance). Such systematic effects can be explored using MA-plots.

Usage

ContrastsPlotter$ma_plot(fc, colour, legend = TRUE, rank = TRUE)

Arguments

fc

fold change abline

colour

column in contrast matrix with colour coding

legend

enable legend default TRUE

rank

default FALSE, if TRUE then rank of avgAbd is used.

Returns

ggplot

Method ma_plotly()

ma plotly

Usage

ContrastsPlotter$ma_plotly(fc, colour, legend = TRUE, rank = FALSE)

Arguments

fc

horizontal lines

colour

column in contrast matrix with colour coding.

legend

enable legend default TRUE

rank

default FALSE, if TRUE then rank of avgAbd is used.

Returns

list of ggplots

Method score_plot()

plot a score against the log2 fc e.g. t-statistic

Usage

ContrastsPlotter$score_plot(scorespec, colour, legend = TRUE)

Arguments

scorespec

list(score="statistics", fcthres = 2, thresh = 5)

colour

column with colour coding

legend

enable legend default TRUE

Returns

list of ggplots

Method score_plotly()

plot a score against the log2 fc e.g. t-statistic

Usage

ContrastsPlotter$score_plotly(scorespec, colour, legend = TRUE)

Arguments

scorespec

list(score="statistics", fcthres = 2, thresh = 5)

colour

column with colour coding

legend

enable legend default TRUE

Returns

list of ggplots

Method barplot_threshold()

shows the number of significant proteins per contrasts

Usage

ContrastsPlotter$barplot_threshold()

Returns

list which contains ggplots and summary tables

Method clone()

The objects of this class are cloneable with this method.

Usage

ContrastsPlotter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

istar <- prolfqua_data('data_ionstar')$normalized()
istar$config <- old2new(istar$config )
istar_data <- dplyr::filter(istar$data ,protein_Id %in% sample(protein_Id, 100))
modelName <- "Model"
modelFunction <-
  strategy_lmer("transformedIntensity  ~ dilution. + (1 | peptide_Id)",
   model_name = modelName)
pepIntensity <- istar_data
config <- istar$config
config$table$hierarchy_keys_depth()
#> [1] "protein_Id"
mod <- build_model(
 pepIntensity,
 modelFunction,
 modelName = modelName,
 subject_Id = config$table$hierarchy_keys_depth())
#> boundary (singular) fit: see help('isSingular')
#> Joining with `by = join_by(protein_Id)`

 #mod$get_coefficients()

 Contr <- c("dil.b_vs_a" = "dilution.b - dilution.a",
  "dil.e_vs_a" = "dilution.e - dilution.a"
  ,"dil.e_vs_b" = "dilution.e - dilution.b",
  "dil.c_vs_b" = "dilution.c - dilution.b"
 )
contrast <- prolfqua::Contrasts$new(mod,
  Contr)
tmp <- contrast$get_contrasts()
#> determine linear functions:
#> compute contrasts:
#> computing contrasts.
#> Warning: There was 1 warning in `dplyr::mutate()`.
#> ℹ In argument: `contrast = purrr::map(linear_model, contrastfun, linfct =
#>   linfct)`.
#> ℹ In group 6: `protein_Id = "sp|O95361|TRI16_HUMAN~598~O95361"`.
#> Caused by warning in `.f()`:
#> ! Model is rank deficient!
#> Joining with `by = join_by(protein_Id, contrast)`

cp <- ContrastsPlotter$new(tmp ,
 contrast$subject_Id,
volcano = list(list(score = "FDR", thresh = 0.1)),
histogram = list(list(score = "p.value", xlim = c(0,1,0.05)),
                 list(score = "FDR", xlim = c(0,1,0.05))),
score =list(list(score = "statistic",  thresh = 5)))
cp$volcano_plotly()
#> $FDR
#> 

cp <- ContrastsPlotter$new(tmp ,
 contrast$subject_Id,
volcano = list(list(score = "FDR", thresh = 0.1)),
histogram = list(list(score = "p.value", xlim = c(0,1,0.05)),
                 list(score = "FDR", xlim = c(0,1,0.05))),
                 fcthresh = NULL,
score =list(list(score = "statistic", thresh = 5)))
cp$fcthresh
#> NULL
cp$volcano_plotly()
#> $FDR
#> 
p <- cp$score_plot(legend=FALSE)
cp$score_plotly()
#> $statistic
#> 
p <- cp$histogram()
p <- cp$histogram_estimate()

res <- cp$volcano()
respltly <- cp$volcano_plotly()

length(respltly)
#> [1] 1
cp$ma_plot()
cp$ma_plotly(rank=TRUE)

res  <- cp$barplot_threshold()
names(res)
#> [1] "FDR"
cp$histogram_diff()
cp$volcano()
#> $FDR
#>