LFQData R6 class

LFQData R6 class

LFQData R6 class

See also

Public fields

config

AnalysisConfiguration

data

data.frame or tibble matching AnalysisConfiguration.

is_pep

todo

prefix

e.g. "peptide_", "protein_", "compound_"

Methods

Public methods

LFQData$new()

LFQData$get_copy()

LFQData$get_sample()

LFQData$subject_id()

LFQData$is_transformed()

LFQData$remove_small_intensities()

LFQData$filter_proteins_by_peptide_count()

LFQData$omit_NA_hierarchy()

LFQData$complete_cases()

LFQData$to_wide()

LFQData$factors()

LFQData$hierarchy_counts()

LFQData$summarize_hierarchy()

LFQData$get_Plotter()

LFQData$get_Writer()

LFQData$get_Summariser()

LFQData$get_Stats()

LFQData$get_Transformer()

LFQData$get_Aggregator()

LFQData$filter_difference()

LFQData$clone()

Method new()

initialize

Usage

LFQData$new(data, config, is_pep = TRUE, prefix = "ms_")

Arguments

data

data.frame

config

configuration

is_pep

todo

prefix

will be use as output prefix

Method get_copy()

get deep copy

Usage

LFQData$get_copy()

Method get_sample()

samples subset of data

Usage

LFQData$get_sample(size = 100, seed = NULL)

Arguments

size

size of subset default 100

seed

set seed

Method subject_id()

get subject ID columns

Usage

LFQData$subject_id()

Method is_transformed()

is data trasfromed

Usage

LFQData$is_transformed(is_transformed)

Arguments

is_transformed

logical

Returns

logical

Method remove_small_intensities()

some software is reporting NA's as 0, you must remove it from your data

Usage

LFQData$remove_small_intensities(threshold = 4)

Arguments

threshold

default 4.

Returns

self

Method filter_proteins_by_peptide_count()

remove proteins with less than X peptides

Usage

LFQData$filter_proteins_by_peptide_count()

Returns

self

Method omit_NA_hierarchy()

Omit NA from intensities per hierarchy, idea is to use it for normalization For instance if a peptide has a missing value in any of the samples it will be rmoved.

Usage

LFQData$omit_NA_hierarchy()

Returns

LFQData with NA omitted.

Method complete_cases()

some software is reporting NA's as 0, you must remove it from your data

Usage

LFQData$complete_cases()

Arguments

threshold

default 4.

Returns

self

Method to_wide()

converts the data to wide

Usage

LFQData$to_wide(as.matrix = FALSE)

Arguments

as.matrix

return as data.frame or matrix

Returns

list with data, annotation, and configuration

Method factors()

Annotation table.

Usage

LFQData$factors()

Returns

data.frame

Method hierarchy_counts()

number of elements at each level

Usage

LFQData$hierarchy_counts()

Method summarize_hierarchy()

e.g. number of peptides per protein etc

Usage

LFQData$summarize_hierarchy()

Returns

data.frame

Method get_Plotter()

get Plotter

Usage

LFQData$get_Plotter()

Returns

LFQDataPlotter

Method get_Writer()

get Writer

Usage

LFQData$get_Writer(format = "xlsx")

Arguments

format

array of formats to write to supported are xlsx, csv and html

Returns

LFQDataPlotter

Method get_Summariser()

get Summariser

Usage

LFQData$get_Summariser()

Returns

LFQDataSummarizer

Method get_Stats()

get Stats

Usage

LFQData$get_Stats()

Returns

LFQDataStats

Method get_Transformer()

get Stats

Usage

LFQData$get_Transformer()

Returns

LFQDataTransformer

Method get_Aggregator()

get Aggregator

Usage

LFQData$get_Aggregator()

Returns

LFQDataAggregator

Method filter_difference()

get difference of self with other if other is subset of self

Usage

LFQData$filter_difference(other)

Arguments

other

a filtered LFQData set

Details

Use to compare filtering results obtained from self, e.g. which proteins and peptides were removed (other)

Returns

LFQData

Method clone()

The objects of this class are cloneable with this method.

Usage

LFQData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(tidyverse)
library(prolfqua)

istar <- prolfqua::data_ionstar$filtered()

#> Column added : nr_peptide_Id_IN_protein_Id

data <- istar$data %>% dplyr::filter(protein_Id %in% sample(protein_Id, 100))
lfqdata <- LFQData$new(data, istar$config)
tmp <- lfqdata$to_wide()
stopifnot("data.frame" %in% class(tmp$data))
tmp <- lfqdata$to_wide(as.matrix = TRUE)
stopifnot("matrix" %in% class(tmp$data))
lfqdata$factors()

#> # A tibble: 20 x 4
#>    raw.file                                          sampleName dilution. run_Id
#>    <chr>                                             <chr>      <chr>     <chr> 
#>  1 b03_10_150304_human_ecoli_a_3ul_3um_column_95_hc~ a~10       a         10    
#>  2 b03_11_150304_human_ecoli_a_3ul_3um_column_95_hc~ a~11       a         11    
#>  3 b03_20_150304_human_ecoli_a_3ul_3um_column_95_hc~ a~20       a         20    
#>  4 b03_21_150304_human_ecoli_a_3ul_3um_column_95_hc~ a~21       a         21    
#>  5 b03_02_150304_human_ecoli_b_3ul_3um_column_95_hc~ b~02       b         02    
#>  6 b03_09_150304_human_ecoli_b_3ul_3um_column_95_hc~ b~09       b         09    
#>  7 b03_12_150304_human_ecoli_b_3ul_3um_column_95_hc~ b~12       b         12    
#>  8 b03_19_150304_human_ecoli_b_3ul_3um_column_95_hc~ b~19       b         19    
#>  9 b03_03_150304_human_ecoli_c_3ul_3um_column_95_hc~ c~03       c         03    
#> 10 b03_08_150304_human_ecoli_c_3ul_3um_column_95_hc~ c~08       c         08    
#> 11 b03_13_150304_human_ecoli_c_3ul_3um_column_95_hc~ c~13       c         13    
#> 12 b03_18_150304_human_ecoli_c_3ul_3um_column_95_hc~ c~18       c         18    
#> 13 b03_04_150304_human_ecoli_d_3ul_3um_column_95_hc~ d~04       d         04    
#> 14 b03_07_150304_human_ecoli_d_3ul_3um_column_95_hc~ d~07       d         07    
#> 15 b03_14_150304_human_ecoli_d_3ul_3um_column_95_hc~ d~14       d         14    
#> 16 b03_17_150304_human_ecoli_d_3ul_3um_column_95_hc~ d~17       d         17    
#> 17 b03_05_150304_human_ecoli_e_3ul_3um_column_95_hc~ e~05       e         05    
#> 18 b03_06_150304_human_ecoli_e_3ul_3um_column_95_hc~ e~06       e         06    
#> 19 b03_15_150304_human_ecoli_e_3ul_3um_column_95_hc~ e~15       e         15    
#> 20 b03_16_150304_human_ecoli_e_3ul_3um_column_95_hc~ e~16       e         16    

stopifnot(lfqdata$is_transformed()==FALSE)
lfqdata$summarize_hierarchy()

#> # A tibble: 62 x 3
#>    protein_Id                       isotope_n peptide_Id_n
#>    <chr>                                <int>        <int>
#>  1 CON__P02769~18~NA                        1           34
#>  2 CON__P34955~32~NA                        1           13
#>  3 CON__Q3SZR3~50~NA                        1            7
#>  4 sp|O95373|IPO7_HUMAN~600~O95373          1           26
#>  5 sp|P02924|ARAF_ECOLI~705~P02924          1            6
#>  6 sp|P04150|GCR_HUMAN~721~P04150           1            5
#>  7 sp|P07602|SAP_HUMAN~814~P07602           1           12
#>  8 sp|P07951|TPM2_HUMAN~829~P07951          1           21
#>  9 sp|P08238|HS90B_HUMAN~837~P08238         1           41
#> 10 sp|P0A6L2|DAPA_ECOLI~914~P0A6L2          1            5
#> # ... with 52 more rows

lfqdata$omit_NA_hierarchy()

#> completing cases

#> Joining, by = c("protein_Id", "peptide_Id")

#> <LFQData>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     complete_cases: function () 
#>     config: AnalysisConfiguration, R6
#>     data: tbl_df, tbl, data.frame
#>     factors: function () 
#>     filter_difference: function (other) 
#>     filter_proteins_by_peptide_count: function () 
#>     get_Aggregator: function () 
#>     get_copy: function () 
#>     get_Plotter: function () 
#>     get_sample: function (size = 100, seed = NULL) 
#>     get_Stats: function () 
#>     get_Summariser: function () 
#>     get_Transformer: function () 
#>     get_Writer: function (format = "xlsx") 
#>     hierarchy_counts: function () 
#>     initialize: function (data, config, is_pep = TRUE, prefix = "ms_") 
#>     is_pep: TRUE
#>     is_transformed: function (is_transformed) 
#>     omit_NA_hierarchy: function () 
#>     prefix: ms_
#>     remove_small_intensities: function (threshold = 4) 
#>     subject_id: function () 
#>     summarize_hierarchy: function () 
#>     to_wide: function (as.matrix = FALSE) 

stopifnot("LFQData" %in% class(lfqdata$get_copy()))
stopifnot("LFQDataTransformer" %in% class(lfqdata$get_Transformer()))
stopifnot("LFQDataStats" %in% class(lfqdata$get_Stats()))

#> completing cases

stopifnot("LFQDataSummariser" %in% class(lfqdata$get_Summariser()))
stopifnot("LFQDataPlotter" %in% class(lfqdata$get_Plotter()))
stopifnot("LFQDataWriter" %in% class(lfqdata$get_Writer()))
stopifnot("LFQDataAggregator" %in% class(lfqdata$get_Aggregator()))