Create various visualization of the LFQdata

Create various visualization of the LFQdata

Create various visualization of the LFQdata

See also

Public fields

lfq

LFQData object

prefix

prefix to figure names when writing, e.g. protein_

file_paths_pdf

with paths to figures

file_paths_html

with paths to figures

Methods

Public methods

LFQDataPlotter$new()

LFQDataPlotter$heatmap()

LFQDataPlotter$heatmap_cor()

LFQDataPlotter$pca()

LFQDataPlotter$pca_plotly()

LFQDataPlotter$boxplots()

LFQDataPlotter$missigness_histogram()

LFQDataPlotter$missingness_per_condition()

LFQDataPlotter$missingness_per_condition_cumsum()

LFQDataPlotter$NA_heatmap()

LFQDataPlotter$intensity_distribution_density()

LFQDataPlotter$intensity_distribution_violin()

LFQDataPlotter$pairs_smooth()

LFQDataPlotter$sample_correlation()

LFQDataPlotter$write_boxplots()

LFQDataPlotter$write_pltly()

LFQDataPlotter$write_pdf()

LFQDataPlotter$write()

LFQDataPlotter$clone()

Method new()

create LFQDataPlotter

Usage

LFQDataPlotter$new(lfqdata, prefix = "ms_")

Arguments

lfqdata

LFQData

prefix

will be prepended to outputs written

Method heatmap()

heatmap of intensities

Usage

LFQDataPlotter$heatmap(na_fraction = 0.3)

Arguments

na_fraction

max fraction of NA's per row

Returns

pheatmap

Method heatmap_cor()

heatmap of sample correlations

Usage

LFQDataPlotter$heatmap_cor()

Returns

pheatmap

Method pca()

pca plot

Usage

LFQDataPlotter$pca()

Returns

ggplot

Method pca_plotly()

pca plot

Usage

LFQDataPlotter$pca_plotly()

Returns

plotly

Method boxplots()

boxplots for all proteins

Usage

LFQDataPlotter$boxplots()

Returns

tibble with column boxplots containing ggplot objects

Method missigness_histogram()

histogram of intensities given number of missing in conditions

Usage

LFQDataPlotter$missigness_histogram()

Returns

ggplot

Method missingness_per_condition()

barplot with number of features with 1,2, etc missing in condition

Usage

LFQDataPlotter$missingness_per_condition()

Returns

ggplot

Method missingness_per_condition_cumsum()

barplot with cumulative sum of features with 1,2, etc missing in condition

Usage

LFQDataPlotter$missingness_per_condition_cumsum()

Returns

ggplot

Method NA_heatmap()

heatmap of features with missing values

Usage

LFQDataPlotter$NA_heatmap()

Returns

ggplot

Method intensity_distribution_density()

density distribution of intensities

Usage

LFQDataPlotter$intensity_distribution_density()

Returns

ggplot

Method intensity_distribution_violin()

Violinplot showing distribution of intensities in all samples

Usage

LFQDataPlotter$intensity_distribution_violin()

Returns

ggplot

Method pairs_smooth()

pairsplot of intensities

Usage

LFQDataPlotter$pairs_smooth(max = 10)

Arguments

max

maximal number of samples to show

Returns

NULL

Method sample_correlation()

plot of sample correlations

Usage

LFQDataPlotter$sample_correlation()

Returns

NULL

Method write_boxplots()

write boxplots to file

Usage

LFQDataPlotter$write_boxplots(path_qc, width = 6, height = 6)

Arguments

path_qc

path to write to

width

fig width

height

fig height

Method write_pltly()

write pltly figures to path_qc

Usage

LFQDataPlotter$write_pltly(fig, path_qc, fig_name)

Arguments

fig

pltly figure

path_qc

path to write to

fig_name

file name (without extension)

Returns

path the file was written to.

Method write_pdf()

write figure to pdf

Usage

LFQDataPlotter$write_pdf(fig, path_qc, fig_name, width = 7, height = 7)

Arguments

fig

ggplot or pheatmap

path_qc

path to write to

fig_name

name of figure (no extension)

width

figure width

height

figure height

Returns

path the file was written to

Method write()

write heatmaps and pca plots to files

Usage

LFQDataPlotter$write(path_qc)

Arguments

path_qc

path to write to

Method clone()

The objects of this class are cloneable with this method.

Usage

LFQDataPlotter$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(prolfqua)
istar <- prolfqua::dataIonstarProtein_subsetNorm

istar$data <- istar$data %>% dplyr::filter(protein_Id %in% sample(protein_Id, 100))

lfqdata <- LFQData$new(
 istar$data,
 istar$config)
#LFQDataPlotter$debug("pca_plotly")
lfqplotter <- lfqdata$get_Plotter()
graphics.off()
lfqplotter$heatmap()
graphics.off()
lfqplotter$heatmap_cor()
graphics.off()
lfqplotter$pca()

#> Joining, by = "sampleName"

lfqplotter$pca_plotly()

#> Joining, by = "sampleName"

#tmp <- lfqplotter$boxplots()
#tmp$boxplot[[1]]
lfqplotter$missigness_histogram()

#> completing cases

#> isotope ~ dilution.

#> `stat_bin()` using `bins = 30`. Pick better value with `binwidth`.

lfqplotter$missingness_per_condition()

#> completing cases

lfqplotter$missingness_per_condition_cumsum()

#> completing cases

#> isotope ~ dilution.

lfqplotter$NA_heatmap()

#> rows with NA's: 17; all rows :82

lfqplotter$intensity_distribution_density()

#> Warning: Removed 72 rows containing non-finite values (stat_density).

lfqplotter$intensity_distribution_violin()

#> Warning: Removed 72 rows containing non-finite values (stat_ydensity).

lfqplotter$pairs_smooth()

#> NULL

lfqplotter$sample_correlation()
prolfqua::plot_sample_correlation(istar$data, istar$config)