compute group mean by LOD

compute group mean by LOD

compute group mean by LOD

Details

weight lod by nr of NA's $(LOD * nrNas + meanAbundance *nrObs)/(nrMeasured)$

Public fields

data

data

config

config

prob

quantile of groups with one observed value to estimate LOD

keep

stats which might be time consuming to compute

weighted

should we weight the LOD

Active bindings

keep

stats which might be time consuming to compute

Methods

Public methods

MissingHelpers$new()

MissingHelpers$get_stats()

MissingHelpers$get_LOD()

MissingHelpers$impute_weighted_lod()

MissingHelpers$impute_lod()

MissingHelpers$get_poolvar()

MissingHelpers$get_contrast_estimates()

MissingHelpers$get_contrasts()

MissingHelpers$clone()

Method new()

initialize

Usage

MissingHelpers$new(data, config, prob = 0.5, weighted = TRUE)

Arguments

data

data

config

config

prob

default 0.5, median of groups with one observed value

weighted

should group average be computed used weighting, default TRUE.

Method get_stats()

Usage

MissingHelpers$get_stats()

Method get_LOD()

Usage

MissingHelpers$get_LOD()

Method impute_weighted_lod()

Usage

MissingHelpers$impute_weighted_lod()

Method impute_lod()

Usage

MissingHelpers$impute_lod()

Method get_poolvar()

Usage

MissingHelpers$get_poolvar(prob = 0.75)

Method get_contrast_estimates()

get contrast estimates

Usage

MissingHelpers$get_contrast_estimates(Contrasts)

Arguments

Contrasts

named array with contrasts

Method get_contrasts()

Usage

MissingHelpers$get_contrasts(Contrasts, confint = 0.95, all = FALSE)

Method clone()

The objects of this class are cloneable with this method.

Usage

MissingHelpers$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

Contrasts <- c("group.b-a" = "group_A - group_B", "group.a-ctrl" = "group_A - group_Ctrl")
dd <- prolfqua::sim_lfq_data_protein_config(Nprot = 100,weight_missing = 2)
#> creating sampleName from fileName column
#> completing cases
mh <- prolfqua::MissingHelpers$new(dd$data, dd$config, prob = 0.8,weighted = TRUE)
mh$get_stats()
#> [1] "group_"
#> completing cases
#> # A tibble: 300 × 11
#>    protein_Id  group_ nrReplicates nrMeasured nrNAs     sd     var meanAbundance
#>    <chr>       <chr>         <int>      <int> <int>  <dbl>   <dbl>         <dbl>
#>  1 0EfVhX~3967 A                 4          4     0  1.43   2.05            22.0
#>  2 0EfVhX~3967 B                 4          4     0  0.965  0.930           19.2
#>  3 0EfVhX~3967 Ctrl              4          1     3 NA     NA               20.6
#>  4 0YSKpy~2865 A                 4          1     3 NA     NA               18.2
#>  5 0YSKpy~2865 B                 4          0     4 NA     NA              NaN  
#>  6 0YSKpy~2865 Ctrl              4          2     2  1.55   2.41            18.5
#>  7 0m5WN4~6025 A                 4          1     3 NA     NA               17.9
#>  8 0m5WN4~6025 B                 4          3     1  0.615  0.378           24.1
#>  9 0m5WN4~6025 Ctrl              4          3     1  0.145  0.0211          20.2
#> 10 3QLHfm~8938 A                 4          3     1  1.05   1.10            22.9
#> # ℹ 290 more rows
#> # ℹ 3 more variables: medianAbundance <dbl>, CV <dbl>, interaction <fct>
mh$get_LOD()
#>      80% 
#> 19.97677 
mh$impute_weighted_lod()
#> # A tibble: 300 × 13
#>    protein_Id  group_ nrReplicates nrMeasured nrNAs     sd     var meanAbundance
#>    <chr>       <chr>         <int>      <int> <int>  <dbl>   <dbl>         <dbl>
#>  1 0EfVhX~3967 A                 4          4     0  1.43   2.05            22.0
#>  2 0EfVhX~3967 B                 4          4     0  0.965  0.930           19.2
#>  3 0EfVhX~3967 Ctrl              4          1     3 NA     NA               20.6
#>  4 0YSKpy~2865 A                 4          1     3 NA     NA               18.2
#>  5 0YSKpy~2865 B                 4          0     4 NA     NA              NaN  
#>  6 0YSKpy~2865 Ctrl              4          2     2  1.55   2.41            18.5
#>  7 0m5WN4~6025 A                 4          1     3 NA     NA               17.9
#>  8 0m5WN4~6025 B                 4          3     1  0.615  0.378           24.1
#>  9 0m5WN4~6025 Ctrl              4          3     1  0.145  0.0211          20.2
#> 10 3QLHfm~8938 A                 4          3     1  1.05   1.10            22.9
#> # ℹ 290 more rows
#> # ℹ 5 more variables: medianAbundance <dbl>, CV <dbl>, interaction <fct>,
#> #   meanAbundanceZero <dbl>, meanAbundanceImp <dbl>
mh$impute_lod()
#> # A tibble: 300 × 12
#>    protein_Id  group_ nrReplicates nrMeasured nrNAs     sd     var meanAbundance
#>    <chr>       <chr>         <int>      <int> <int>  <dbl>   <dbl>         <dbl>
#>  1 0EfVhX~3967 A                 4          4     0  1.43   2.05            22.0
#>  2 0EfVhX~3967 B                 4          4     0  0.965  0.930           19.2
#>  3 0EfVhX~3967 Ctrl              4          1     3 NA     NA               20.6
#>  4 0YSKpy~2865 A                 4          1     3 NA     NA               18.2
#>  5 0YSKpy~2865 B                 4          0     4 NA     NA              NaN  
#>  6 0YSKpy~2865 Ctrl              4          2     2  1.55   2.41            18.5
#>  7 0m5WN4~6025 A                 4          1     3 NA     NA               17.9
#>  8 0m5WN4~6025 B                 4          3     1  0.615  0.378           24.1
#>  9 0m5WN4~6025 Ctrl              4          3     1  0.145  0.0211          20.2
#> 10 3QLHfm~8938 A                 4          3     1  1.05   1.10            22.9
#> # ℹ 290 more rows
#> # ℹ 4 more variables: medianAbundance <dbl>, CV <dbl>, interaction <fct>,
#> #   meanAbundanceImp <dbl>
mh$get_poolvar()
#> # A tibble: 100 × 9
#>    protein_Id  n.groups     n    df    sd   sdT  mean meanAll nrMeasured
#>    <chr>          <int> <int> <dbl> <dbl> <dbl> <dbl>   <dbl>      <int>
#>  1 0EfVhX~3967        2     8     6 1.22  0.864  20.6    20.6          9
#>  2 0YSKpy~2865        1     2     1 1.55  1.55   18.5    18.4          3
#>  3 0m5WN4~6025        2     6     4 0.447 0.365  22.1    21.5          7
#>  4 3QLHfm~8938        3    10     7 0.923 0.715  22.6    22.6         10
#>  5 3QYop0~7543        3    10     7 0.808 0.626  22.9    22.9         10
#>  6 76k03k~7094        3    11     8 1.09  0.803  24.8    24.8         11
#>  7 7QuTub~1867        2     8     6 0.644 0.455  23.3    23.2          9
#>  8 7cbcrd~7351        3    12     9 0.889 0.629  24.9    24.9         12
#>  9 7soopj~5352        2     5     3 0.920 0.823  20.0    20.0          5
#> 10 7zeekV~7127        3     9     6 1.06  0.864  20.9    20.9          9
#> # ℹ 90 more rows
bb <- mh$get_contrast_estimates(Contrasts)
#> group.b-a=group_A - group_B
#> group.a-ctrl=group_A - group_Ctrl
#> group.b-a=group_A - group_B
#> group.a-ctrl=group_A - group_Ctrl
#> group.b-a=group_A - group_B
#> group.a-ctrl=group_A - group_Ctrl
mh$get_contrasts(Contrasts)
#> group.b-a=group_A - group_B
#> group.a-ctrl=group_A - group_Ctrl
#> group.b-a=group_A - group_B
#> group.a-ctrl=group_A - group_Ctrl
#> group.b-a=group_A - group_B
#> group.a-ctrl=group_A - group_Ctrl
#> # A tibble: 200 × 18
#>    protein_Id  meanAbundanceImp_group_1 meanAbundanceImp_group_2 estimate
#>    <chr>                          <dbl>                    <dbl>    <dbl>
#>  1 0EfVhX~3967                     22.0                     19.2    2.84 
#>  2 0YSKpy~2865                     19.5                     20.0    0    
#>  3 0m5WN4~6025                     19.5                     23.1   -3.59 
#>  4 3QLHfm~8938                     22.2                     21.8    0.365
#>  5 3QYop0~7543                     22.5                     22.1    0.433
#>  6 76k03k~7094                     25.3                     24.3    1.02 
#>  7 7QuTub~1867                     20.4                     24.6   -4.12 
#>  8 7cbcrd~7351                     26.7                     23.6    3.17 
#>  9 7soopj~5352                     20.1                     20.0    0.113
#> 10 7zeekV~7127                     20.6                     20.8   -0.196
#> # ℹ 190 more rows
#> # ℹ 14 more variables: group_1_name <chr>, group_2_name <chr>, contrast <chr>,
#> #   avgAbd <dbl>, indic <dbl>, nrMeasured_group_1 <int>,
#> #   nrMeasured_group_2 <int>, df <dbl>, sd <dbl>, sdT <dbl>, statistic <dbl>,
#> #   p.value <dbl>, conf.low <dbl>, conf.high <dbl>

dd <- prolfqua::sim_lfq_data_protein_2Factor_config(Nprot = 100,weight_missing = 0.1)
#> creating sampleName from fileName column
#> completing cases

Contrasts <- c("c1" = "TreatmentA - TreatmentB",
               "C2" = "BackgroundX- BackgroundZ",
               "c3" = "`TreatmentA:BackgroundX` - `TreatmentA:BackgroundZ`",
               "c4" = "`TreatmentB:BackgroundX` - `TreatmentB:BackgroundZ`"
               )
mh <- prolfqua::MissingHelpers$new(dd$data, dd$config, prob = 0.8,weighted = TRUE)
mh$get_stats()$interaction |> table()
#> [1] "Treatment"  "Background"
#> completing cases
#> [1] "Treatment"
#> completing cases
#> [1] "Background"
#> completing cases
#> 
#> TreatmentA:BackgroundX TreatmentB:BackgroundX TreatmentA:BackgroundZ 
#>                    100                    100                    100 
#> TreatmentB:BackgroundZ             TreatmentA             TreatmentB 
#>                    100                    100                    100 
#>            BackgroundX            BackgroundZ 
#>                    100                    100 
mh$get_contrast_estimates(Contrasts)
#> c1=TreatmentA - TreatmentB
#> C2=BackgroundX- BackgroundZ
#> c3=`TreatmentA:BackgroundX` - `TreatmentA:BackgroundZ`
#> c4=`TreatmentB:BackgroundX` - `TreatmentB:BackgroundZ`
#> c1=TreatmentA - TreatmentB
#> C2=BackgroundX- BackgroundZ
#> c3=`TreatmentA:BackgroundX` - `TreatmentA:BackgroundZ`
#> c4=`TreatmentB:BackgroundX` - `TreatmentB:BackgroundZ`
#> c1=TreatmentA - TreatmentB
#> C2=BackgroundX- BackgroundZ
#> c3=`TreatmentA:BackgroundX` - `TreatmentA:BackgroundZ`
#> c4=`TreatmentB:BackgroundX` - `TreatmentB:BackgroundZ`
#> # A tibble: 400 × 11
#>    protein_Id  meanAbundanceImp_group_1 meanAbundanceImp_group_2 estimate
#>    <chr>                          <dbl>                    <dbl>    <dbl>
#>  1 0EfVhX~3967                     21.1                     20.5    0.608
#>  2 0YSKpy~2865                     18.3                     17.1    1.15 
#>  3 0m5WN4~6025                     18.1                     22.2   -4.09 
#>  4 3QLHfm~8938                     23.2                     22.5    0.712
#>  5 3QYop0~7543                     23.3                     23.7   -0.310
#>  6 76k03k~7094                     24.3                     24.6   -0.343
#>  7 7QuTub~1867                     21.6                     25.7   -4.14 
#>  8 7cbcrd~7351                     25.6                     24.8    0.804
#>  9 7soopj~5352                     19.7                     21.6   -1.90 
#> 10 7zeekV~7127                     21.1                     20.3    0.720
#> # ℹ 390 more rows
#> # ℹ 7 more variables: group_1_name <chr>, group_2_name <chr>, contrast <chr>,
#> #   avgAbd <dbl>, indic <dbl>, nrMeasured_group_1 <int>,
#> #   nrMeasured_group_2 <int>