Compute theoretical sample sizes from factor level standard deviations

Compute theoretical sample sizes from factor level standard deviations

lfq_power_t_test_proteins(
  stats_res,
  delta = c(0.59, 1, 2),
  power = 0.8,
  sig.level = 0.05,
  min.n = 1.5
)

Arguments

stats_res
data.frame `summarize_cv` output

delta
effect size you are interested in

power
of test

min.n
smallest n to determine

sigma.level
P-Value

Examples

library(prolfqua)
library(tidyverse)
bb1 <- prolfqua::ionstar$normalized()
config <- bb1$config$clone( deep = TRUE)
data2 <- bb1$data
stats_res <- summarize_cv(data2, config, all = FALSE)
bb <- lfq_power_t_test_proteins(stats_res)
head(bb)

#> # A tibble: 6 x 10
#>   protein_Id  peptide_Id dilution.     n not_na    sd   mean delta N_exact     N
#>   <chr>       <chr>      <chr>     <int>  <int> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1 CON__P0103~ AELADQAAS~ e             4      2 0.647 -2.21   0.59   19.9     20
#> 2 CON__P0103~ ILSLAQDQV~ a             4      4 0.499 -0.973  0.59   12.3     13
#> 3 CON__P0103~ ILSLAQDQV~ b             4      4 0.544 -1.03   0.59   14.4     15
#> 4 CON__P0103~ ILSLAQDQV~ c             4      4 0.384 -1.15   0.59    7.74     8
#> 5 CON__P0103~ ILSLAQDQV~ d             4      4 0.391 -1.41   0.59    7.97     8
#> 6 CON__P0103~ ILSLAQDQV~ e             4      4 0.403 -1.35   0.59    8.40     9