Compute theoretical sample sizes from factor level standard deviations

Compute theoretical sample sizes from factor level standard deviations

lfq_power_t_test_proteins(
  stats_res,
  delta = c(0.59, 1, 2),
  power = 0.8,
  sig.level = 0.05,
  min.n = 1.5
)

Arguments

stats_res
data.frame `summarize_stats` output

delta
effect size you are interested in

power
of test

min.n
smallest n to determine

sigma.level
P-Value

See also

Examples

library(prolfqua)
library(tidyverse)
bb1 <- prolfqua::data_ionstar$normalized()
config <- bb1$config$clone( deep = TRUE)
data2 <- bb1$data
stats_res <- summarize_stats(data2, config)

#> completing cases

bb <- lfq_power_t_test_proteins(stats_res)
head(bb)

#> # A tibble: 6 x 11
#>   protein_Id  peptide_Id dilution.     n not_na    sd   var   mean delta N_exact
#>   <chr>       <chr>      <chr>     <int>  <int> <dbl> <dbl>  <dbl> <dbl>   <dbl>
#> 1 CON__P0103~ AELADQAAS~ e             4      2 0.647 0.419 -2.21   0.59   19.9 
#> 2 CON__P0103~ ILSLAQDQV~ a             4      4 0.499 0.249 -0.973  0.59   12.3 
#> 3 CON__P0103~ ILSLAQDQV~ b             4      4 0.544 0.296 -1.03   0.59   14.4 
#> 4 CON__P0103~ ILSLAQDQV~ c             4      4 0.384 0.148 -1.15   0.59    7.74
#> 5 CON__P0103~ ILSLAQDQV~ d             4      4 0.391 0.153 -1.41   0.59    7.97
#> 6 CON__P0103~ ILSLAQDQV~ e             4      4 0.403 0.162 -1.35   0.59    8.40
#> # ... with 1 more variable: N <dbl>