Compute theoretical sample sizes from factor level standard deviations

Compute theoretical sample sizes from factor level standard deviations

lfq_power_t_test_quantiles(
  pdata,
  config,
  delta = 1,
  power = 0.8,
  sig.level = 0.05,
  probs = seq(0.5, 0.9, by = 0.1)
)

Arguments

pdata

data.frame

config

AnalysisConfiguration

delta

effect size you are interested in

power

of test

sigma.level

P-Value

See also

Examples

bb1 <- prolfqua::sim_lfq_data_peptide_config()
#> creating sampleName from fileName column
#> Warning: no nr_children column specified in the data, adding column nr_children and setting to 1.
#> completing cases
config <- bb1$config
data2 <- bb1$data

res <- lfq_power_t_test_quantiles(data2, config)
#> Warning: Intensities are not transformed yet.
#> [1] "group_"
#> completing cases
res$summary
#>   quantile probs        sd FC=2
#> 1      50%   0.5 0.8606778   13
#> 2      60%   0.6 1.0041074   17
#> 3      70%   0.7 1.0808016   20
#> 4      80%   0.8 1.2819172   27
#> 5      90%   0.9 1.4827429   36
stats_res <- summarize_stats(data2, config)
#> [1] "group_"
#> completing cases
res <- lfq_power_t_test_quantiles(data2, config, delta = 2)
#> Warning: Intensities are not transformed yet.
#> [1] "group_"
#> completing cases
res <- lfq_power_t_test_quantiles(data2, config, delta = c(0.5,1,2))
#> Warning: Intensities are not transformed yet.
#> [1] "group_"
#> completing cases