Marks peptides which do not correlate with other peptides of a protein

It computes the pairwise correlation among all peptides and marks those with with a corralation less then minCorrelation to all other peptides

mark_decorelated(data, config, minCorrelation = 0.7)

Arguments

data

data

config

configuration

minCorrelation

smallest allowed correlation

TODO

do something with warnings of type "the standard deviation is zero".

do investigate In max(x, na.rm = TRUE) : no non-missing arguments to max; returning -Inf

Examples

bb <- prolfqua_data('data_ionstar')$filtered()
#> Column added : nr_peptide_Id_IN_protein_Id
stopifnot( nrow(bb$data) == 25780)
conf <- old2new(bb$config$clone(deep=TRUE))
data <- bb$data |> dplyr::ungroup()
dim(data)
#> [1] 25780    10
dataI <- mark_decorelated(data, conf)
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 7: `row.names = "ASVHTLSGHTNAVATVR"`, `col.names =
#>   "VYREDDTATEETHPVSWKPEIIK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "DVLAHQVPNAK"`, `col.names =
#>   "LHTKPSQAPAVEVAPAGASYNPSFEDHQTLLSAAHEVELQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 9 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "AQLSQYFDAVK"`, `col.names = "FVPTALR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 8 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 75: `row.names = "NCPHVVVGTPGR"`, `col.names =
#>   "VNIVFNYDMPEDSDTYLHR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 60: `row.names = "IPELDMTEVVAPFMANIPLLLYPQDGPR"`, `col.names =
#>   "TEKEILDAFDK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 10 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 31: `row.names = "FEAAETLEEAAMR"`, `col.names = "YEEVEYYYQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 9 remaining warnings.
#> Warning: There were 14 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 5: `row.names = "AIPNQGEILVIR"`, `col.names = "IEGSGDQVDTLELSGGAR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 13 remaining warnings.
#> Warning: There were 268 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 8: `row.names = "AFQQNLVSSALFSK"`, `col.names =
#>   "HHVLQNDVLAHQSTVEAVNK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 267 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 30: `row.names = "GDEVIFGEFSWPK"`, `col.names = "SAPLEIGLQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 33 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "DFNPTATVK"`, `col.names = "EGLLFEGR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 32 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 94: `row.names = "KAEAQIAAK"`, `col.names =
#>   "MNLGVGAYRDDNGKPYVLPSVR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 67 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 29: `row.names = "AEFVEVTK"`, `col.names = "VHKECCHGDLLECADDR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 66 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "DQQNLPYGVTPASPSGHSQGR"`, `col.names =
#>   "TTPTPSTNSVLSTSTNR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "LLDGFFIRPFYK"`, `col.names =
#>   "RPTPQDNLTDAENGNIQLQAQR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 8: `row.names = "GVVVQDAALLESGAAIR"`, `col.names =
#>   "TAEGGAEHVQPITGDNIVNVLDDFR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 19 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 9: `row.names = "ANKLDHVVTIIK"`, `col.names = "VEDLTFTSPFCLQVK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 18 remaining warnings.
#> Warning: There were 5 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "FALEVPGLR"`, `col.names = "LSETSPAFR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 4 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "MTEWETAAPAVAETPDIK"`, `col.names = "QAVDVSPLRR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 33 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "AIEQLKEQAATGK"`, `col.names = "ATIFNLK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 32 remaining warnings.
#> Warning: There were 6 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AAIQSQLDGVR"`, `col.names = "KYVLDDLIVAK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 5 remaining warnings.
#> Warning: There were 11 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "ELQHEKDTSSLLR"`, `col.names = "ICSLQEEQQQLR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 10 remaining warnings.
#> Warning: There were 8 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "AGQMVTVWAAGAGVAHSPPSTLVWK"`, `col.names =
#>   "ELEEAMAGERDK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 7 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AELDLMAPGPLPR"`, `col.names = "VTNTTPWEHPYPDLPSR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 387: `row.names = "GPLEYVPSAK"`, `col.names = "TAVFGFETSEAK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "EVAWEQNYPDR"`, `col.names = "ITGIDSSPAMIAEAR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 6 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 220: `row.names = "VLPLEEAYR"`, `col.names = "VYIGSFWAQPLQNTDNRR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 5 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 121: `row.names = "HIAEDSDRKYEEVAR"`, `col.names = "MELQEMQLK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 7: `row.names = "AQEAPGQAEPPAAAEVQGAGNENEPR"`, `col.names =
#>   "LLVVDRETDEFFK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AQYVLAEQVTR"`, `col.names =
#>   "FQQAGHKPVALVGGATGLIGDPSFK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 4: `row.names = "ASAGSYISALR"`, `col.names = "GPGLYYVDEHGTR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 146: `row.names = "ETAPGDISPYTIPEEDR"`, `col.names = "PVMQALLHIIR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "LKTEEGEIDYSAEEGENRR"`, `col.names =
#>   "TIPGTALVEMGDEYAVER"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 6 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 27: `row.names = "FSLFAERK"`, `col.names =
#>   "QLSCVANQNGSQADCELGNPFKR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 5 remaining warnings.
#> Warning: There were 10 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "AFAASNELEVNATHKDR"`, `col.names =
#>   "EGITATDLVLTVTQMLR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 9 remaining warnings.
#> Warning: There were 7 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 1: `row.names = "AIQAELLK"`, `col.names = "AIQAELLK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 6 remaining warnings.
#> Warning: There were 16 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 52: `row.names = "EEMLQMIR"`, `col.names = "KANYAVDAYFR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 15 remaining warnings.
#> Warning: There were 14 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 577: `row.names = "DHRDNLEFFLAGIGR"`, `col.names =
#>   "FCFTPHTEEGCLSER"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 13 remaining warnings.
#> Warning: There were 3 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 1: `row.names = "FSIALLNQAVER"`, `col.names = "FSIALLNQAVER"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 2 remaining warnings.
#> Warning: There were 8 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 5: `row.names = "DFYLATSPPDSFLDDHHLTRPHPER"`, `col.names =
#>   "NHQDPLAVAYHLIIDNR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 7 remaining warnings.
#> Warning: There were 3 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 2: `row.names = "FIDTSQFILNR"`, `col.names = "SKPVFSESLSD"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 2 remaining warnings.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 335: `row.names = "EKLCYVALDFEQEMATAASSSSLEK"`, `col.names =
#>   "HQGVMVGMGQKDSYVGDEAQSK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 20: `row.names = "GPRPPNQQQPIEGTDRVEPK"`, `col.names =
#>   "SQADKPVLAIQVLGTVK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 6: `row.names = "ITYVANVNPGGWAPASVLR"`, `col.names =
#>   "RHGSMVSLVSGASGYSATSTSSFK"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 2 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 3: `row.names = "GLSNLFLSCPIPK"`, `col.names = "LLLLAGVDR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 1 remaining warning.
#> Warning: There were 18 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 37: `row.names = "ALSINTLVK"`, `col.names = "FASTVFPSDHIPSR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 17 remaining warnings.
#> Warning: There were 4 warnings in `summarise()`.
#> The first warning was:
#> ℹ In argument: `jcor = (function (x) ...`.
#> ℹ In group 8: `row.names = "ESTSYIEEHIDR"`, `col.names =
#>   "TQYLQSYGPEHLLTFSNLR"`.
#> Caused by warning in `max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 3 remaining warnings.
#> Warning: `cols` is now required when using `unnest()`.
#> ℹ Please use `cols = c(srmDecor)`.