compute per group averages and impute values should generalize at some stage

compute per group averages and impute values should generalize at some stage

missigness_impute_factors_interactions(
  pdata,
  config,
  probs = 0.03,
  value = c("long", "nrReplicates", "nrMeasured", "meanArea", "imputed"),
  add.prefix = FALSE,
  global = TRUE
)

Arguments

pdata

data.frame

config

AnalysisConfiguration

probs

quantile to take average from (default 0.1)

value

use default

add.prefix

use default

See also

Examples

bb <- prolfqua_data('data_ionstar')$filtered()
#> Column added : nr_peptide_Id_IN_protein_Id
stopifnot(nrow(bb$data) == 25780)
configur <- bb$config
data <- bb$data
xx <- complete_cases(data, configur)
#> completing cases

res <- missigness_impute_factors_interactions(xx, configur)
#> completing cases
head(res)
#> # A tibble: 6 × 11
#>   dilution. protein_Id peptide_Id isotope nrReplicates nrNAs meanArea medianArea
#>   <chr>     <chr>      <chr>      <chr>          <int> <int>    <dbl>      <dbl>
#> 1 a         CON__P010… AELADQAAS… light              4     3   7.98e6    7982100
#> 2 a         CON__P010… ILSLAQDQV… light              4     0   1.32e7   13208600
#> 3 a         CON__P010… LLLFSPSVVR light              4     0   7.16e6    6422100
#> 4 a         CON__P027… AEFVEVTK   light              4     0   5.28e8  589010000
#> 5 a         CON__P027… DDPHACYST… light              4     0   2.18e8  229275000
#> 6 a         CON__P027… DDPHACYST… light              4     0   4.44e7   47249500
#> # … with 3 more variables: nrMeasured <int>, interaction <fct>, imputed <dbl>
res <- missigness_impute_factors_interactions(xx, configur, value = "imputed")
#> completing cases
head(res)
#> # A tibble: 6 × 9
#>   value   protein_Id        peptide_Id  isotope dilution.a dilution.b dilution.c
#>   <chr>   <chr>             <chr>       <chr>        <dbl>      <dbl>      <dbl>
#> 1 imputed CON__P01030~9~NA  AELADQAASW… light      7982100    4949600   2908300.
#> 2 imputed CON__P01030~9~NA  ILSLAQDQVG… light     13181700   12526225  10191650 
#> 3 imputed CON__P01030~9~NA  LLLFSPSVVR  light      7158275    6398750   6205275 
#> 4 imputed CON__P02769~18~NA AEFVEVTK    light    528425000  507305000 471840000 
#> 5 imputed CON__P02769~18~NA DDPHACYSTV… light    217962500  212912500 178570000 
#> 6 imputed CON__P02769~18~NA DDPHACYSTV… light     44412000   42483750  38225250 
#> # … with 2 more variables: dilution.d <dbl>, dilution.e <dbl>
res <- missigness_impute_factors_interactions(xx, configur, value = "nrMeasured")
#> completing cases
long <- missigness_impute_factors_interactions(xx, configur, value = "long")
#> completing cases
long
#> # A tibble: 6,445 × 11
#>    dilution. protein_Id        peptide_Id    isotope nrReplicates nrNAs meanArea
#>    <chr>     <chr>             <chr>         <chr>          <int> <int>    <dbl>
#>  1 a         CON__P01030~9~NA  AELADQAASWLTR light              4     3   7.98e6
#>  2 a         CON__P01030~9~NA  ILSLAQDQVGGS… light              4     0   1.32e7
#>  3 a         CON__P01030~9~NA  LLLFSPSVVR    light              4     0   7.16e6
#>  4 a         CON__P02769~18~NA AEFVEVTK      light              4     0   5.28e8
#>  5 a         CON__P02769~18~NA DDPHACYSTVFDK light              4     0   2.18e8
#>  6 a         CON__P02769~18~NA DDPHACYSTVFD… light              4     0   4.44e7
#>  7 a         CON__P02769~18~NA ECCHGDLLECAD… light              4     0   1.21e7
#>  8 a         CON__P02769~18~NA ETYGDMADCCEK  light              4     0   2.49e7
#>  9 a         CON__P02769~18~NA EYEATLEECCAK  light              4     0   2.69e7
#> 10 a         CON__P02769~18~NA FKDLGEEHFK    light              4     0   1.27e9
#> # … with 6,435 more rows, and 4 more variables: medianArea <dbl>,
#> #   nrMeasured <int>, interaction <fct>, imputed <dbl>