Create custom lmer model

The strategy contains functions to fit the model but also compute the contrasts etc.

The strategy contains functions to fit the model but also compute the contrasts etc.

strategy_lmer(
  modelstr,
  model_name = "Model",
  report_columns = c("statistic", "p.value", "p.value.adjusted", "moderated.p.value",
    "moderated.p.value.adjusted")
)

strategy_lm(
  modelstr,
  model_name = "Model",
  report_columns = c("statistic", "p.value", "p.value.adjusted", "moderated.p.value",
    "moderated.p.value.adjusted")
)

strategy_rlm(
  modelstr,
  model_name = "Model",
  report_columns = c("statistic", "p.value", "p.value.adjusted", "moderated.p.value",
    "moderated.p.value.adjusted")
)

strategy_glm(
  modelstr,
  model_name = "Model",
  report_columns = c("statistic", "p.value", "p.value.adjusted", "moderated.p.value",
    "moderated.p.value.adjusted")
)

Arguments

modelstr

model formula

model_name

name of model

report_columns

columns to report

Value

list with model function, contrast computation function etc.

list with model function, contrast computation function etc.

See also

Examples

tmp <- strategy_lmer("Intensity ~ condition + (1|peptide_Id)", model_name = "random_example")
tmp$model_fun(get_formula = TRUE)
#> Intensity ~ condition + (1 | peptide_Id)
#> <environment: 0x2abd273c0>
tmp$isSingular
#> function (x, tol = 1e-04) 
#> {
#>     lwr <- getME(x, "lower")
#>     theta <- getME(x, "theta")
#>     any(theta[lwr == 0] < tol)
#> }
#> <bytecode: 0x159433488>
#> <environment: namespace:lme4>

tmp <- strategy_lm("Intensity ~ condition", model_name = "parallel design")
tmp$model_fun(get_formula = TRUE)
#> Intensity ~ condition
#> <environment: 0x323811c80>
tmp$isSingular
#> function(m){
#>   anyNA <- any(is.na(coefficients(m)))
#>   if (anyNA) {
#>     return(TRUE)
#>   } else {
#>     if (df.residual(m) >= 2) {
#>       return(FALSE)
#>     }
#>     return(TRUE)
#>   }
#> }
#> <environment: namespace:prolfqua>
tmp <- strategy_rlm("Intensity ~ condition", model_name = "parallel design")
tmp$model_fun(get_formula = TRUE)
#> Intensity ~ condition
#> <environment: 0x2fdbafb68>
tmp$isSingular
#> function(m){
#>   anyNA <- any(is.na(coefficients(m)))
#>   if (anyNA) {
#>     return(TRUE)
#>   } else {
#>     if (df.residual(m) >= 2) {
#>       return(FALSE)
#>     }
#>     return(TRUE)
#>   }
#> }
#> <environment: namespace:prolfqua>
tmp <- strategy_glm("Intensity ~ condition", model_name = "parallel design")
tmp$model_fun(get_formula = TRUE)
#> Intensity ~ condition
#> <environment: 0x2c95bbfe0>
tmp$isSingular
#> function(m){
#>   anyNA <- any(is.na(coefficients(m)))
#>   if (anyNA) {
#>     return(TRUE)
#>   } else {
#>     if (df.residual(m) >= 2) {
#>       return(FALSE)
#>     }
#>     return(TRUE)
#>   }
#> }
#> <environment: namespace:prolfqua>